implement & document Target.fetch & Target.xhr
