+++ docs/api/RequestInterception/index.html
@@ -370,7 +370,8 @@
 <dl class="py property">
 <dt class="sig sig-object py" id="selenium_driverless.scripts.network_interceptor.InterceptedRequest.body">
 <em class="property"><span class="pre">property</span><span class="w"> </span></em><span class="sig-name descname"><span class="pre">body</span></span><em class="property"><span class="p"><span class="pre">:</span></span><span class="w"> </span><span class="pre">bytes</span><span class="w"> </span><span class="p"><span class="pre">|</span></span><span class="w"> </span><span class="pre">None</span></em><a class="headerlink" href="#selenium_driverless.scripts.network_interceptor.InterceptedRequest.body" title="Link to this definition"></a></dt>
-<dd><p>get the response body for a request</p>
+<dd><p><strong>async</strong>
+get the response body for a request</p>
 <div class="admonition note">
 <p class="admonition-title">Note</p>
 <p>this uses network and might take some time</p>
