+++ docs/genindex/index.html
@@ -394,6 +394,14 @@
   <td style="width: 33%; vertical-align: top;"><ul>
       <li><a href="../api/RequestInterception/#selenium_driverless.scripts.network_interceptor.InterceptedRequest.fail_request">fail_request() (selenium_driverless.scripts.network_interceptor.InterceptedRequest method)</a>
 </li>
+      <li><a href="../api/Context/#selenium_driverless.types.context.Context.fetch">fetch() (selenium_driverless.types.context.Context method)</a>
+
+      <ul>
+        <li><a href="../api/Target/#selenium_driverless.types.target.Target.fetch">(selenium_driverless.types.target.Target method)</a>
+</li>
+        <li><a href="../api/Chrome/#selenium_driverless.webdriver.Chrome.fetch">(selenium_driverless.webdriver.Chrome method)</a>
+</li>
+      </ul></li>
       <li><a href="../api/Context/#selenium_driverless.types.context.Context.find_element">find_element() (selenium_driverless.types.context.Context method)</a>
 
       <ul>
@@ -414,14 +422,14 @@
         <li><a href="../api/Chrome/#selenium_driverless.webdriver.Chrome.find_elements">(selenium_driverless.webdriver.Chrome method)</a>
 </li>
       </ul></li>
+  </ul></td>
+  <td style="width: 33%; vertical-align: top;"><ul>
       <li><a href="../api/WebELement/#selenium_driverless.types.webelement.WebElement.focus">focus() (selenium_driverless.types.webelement.WebElement method)</a>
 
       <ul>
         <li><a href="../api/Chrome/#selenium_driverless.webdriver.Chrome.focus">(selenium_driverless.webdriver.Chrome method)</a>
 </li>
       </ul></li>
-  </ul></td>
-  <td style="width: 33%; vertical-align: top;"><ul>
       <li><a href="../api/Context/#selenium_driverless.types.context.Context.forward">forward() (selenium_driverless.types.context.Context method)</a>
 
       <ul>
@@ -1133,6 +1141,16 @@
 
 <h2 id="X">X</h2>
 <table style="width: 100%" class="indextable genindextable"><tr>
+  <td style="width: 33%; vertical-align: top;"><ul>
+      <li><a href="../api/Context/#selenium_driverless.types.context.Context.xhr">xhr() (selenium_driverless.types.context.Context method)</a>
+
+      <ul>
+        <li><a href="../api/Target/#selenium_driverless.types.target.Target.xhr">(selenium_driverless.types.target.Target method)</a>
+</li>
+        <li><a href="../api/Chrome/#selenium_driverless.webdriver.Chrome.xhr">(selenium_driverless.webdriver.Chrome method)</a>
+</li>
+      </ul></li>
+  </ul></td>
   <td style="width: 33%; vertical-align: top;"><ul>
       <li><a href="../api/By/#selenium_driverless.types.by.By.XPATH">XPATH (selenium_driverless.types.by.By attribute)</a>
 </li>
