+++ docs/searchindex.js
@@ -1,1 +1,1 @@
-Search.setIndex({"docnames": ["api/By", "api/Chrome", "api/ChromeOptions", "api/Context", "api/RequestInterception", "api/Target", "api/WebELement", "index"], "filenames": ["api/By.rst", "api/Chrome.rst", "api/ChromeOptions.rst", "api/Context.rst", "api/RequestInterception.rst", "api/Target.rst", "api/WebELement.rst", "index.rst"], "titles": ["By Element Locator", "webdriver.Chrome", "ChromeOptions", "Context", "Request-Interception", "Target", "WebElement", "Documentation of Driverless"], "terms": {"class": [0, 1, 2, 3, 4, 5, 6], "selenium_driverless": [0, 1, 2, 3, 4, 5, 6, 7], "type": [0, 1, 2, 3, 4, 5, 6, 7], "sourc": [0, 1, 2, 3, 4, 5, 6], "set": [0, 1, 2, 3, 4, 5, 6], "support": [0, 1, 2, 3, 5, 6], "strategi": 0, "id": [0, 1, 4, 5, 6], "name": [0, 1, 2, 3, 4, 5, 6], "xpath": [0, 3, 5, 7], "tag_nam": [0, 1, 3, 6], "tag": 0, "class_nam": [0, 6], "css_selector": 0, "css": 0, "selector": [0, 3, 5], "alia": [0, 1, 6], "option": [1, 2, 3, 4, 5, 6, 7], "none": [1, 2, 3, 4, 5, 6], "timeout": [1, 2, 3, 4, 5, 6, 7], "30": [1, 3, 4, 5, 6], "debug": 1, "fals": [1, 2, 3, 4, 5, 6], "max_ws_siz": [1, 3, 4, 5], "134217728": 1, "control": 1, "chromium": [1, 2, 3, 5], "base": [1, 3, 5], "browser": [1, 2, 3, 4, 5], "without": [1, 3, 4, 5], "ani": [1, 3, 4, 5, 6], "driver": [1, 2, 3, 4, 5, 7], "properti": [1, 2, 3, 4, 5, 6], "frame_tre": 1, "dict": [1, 2, 3, 4, 5, 6], "async": [1, 2, 3, 4, 5, 6, 7], "all": [1, 2, 3, 5, 6, 7], "nest": [1, 6], "frame": [1, 3, 4, 5], "within": [1, 3], "current": [1, 3, 5, 6], "target": [1, 3, 4, 6, 7], "str": [1, 2, 3, 4, 5, 6], "targetinfo": [1, 3, 5, 7], "context": [1, 5, 7], "incognito": [1, 5], "new_context": 1, "proxy_bypass_list": 1, "proxy_serv": 1, "true": [1, 2, 3, 4, 5, 6, 7], "universal_access_origin": 1, "url": [1, 2, 3, 4, 5, 6], "about": [1, 2], "blank": [1, 2], "creat": [1, 2, 3, 4], "new": [1, 3, 4], "paramet": [1, 2, 3, 4, 5, 6], "first": [1, 2], "tab": [1, 2, 3, 5], "start": [1, 2, 3, 4, 5], "default": [1, 2, 3, 4, 5], "list": [1, 2, 3, 4, 5, 6], "an": [1, 2, 3, 4, 5, 6], "origin": [1, 3, 4, 5], "grant": 1, "unlimit": 1, "cross": [1, 4], "access": [1, 5], "part": 1, "other": [1, 3, 4], "than": 1, "those": [1, 5], "constitut": 1, "ar": [1, 5], "ignor": [1, 2], "return": [1, 2, 3, 4, 5, 6], "The": [1, 2, 3, 4, 5, 6], "proxi": [1, 2, 4], "doesn": [1, 2, 3, 4, 6], "t": [1, 2, 3, 4, 6, 7], "work": 1, "window": [1, 3, 5, 6], "due": [1, 2], "crbug": [1, 2], "1310057": 1, "server": [1, 4], "us": [1, 2, 3, 4, 5, 6], "get_target": 1, "_type": 1, "context_id": [1, 3, 5, 6], "self": 1, "get": [1, 2, 3, 4, 5, 6, 7], "liter": [1, 3, 4, 6], "page": [1, 3, 4, 5, 6, 7], "background_pag": 1, "service_work": 1, "param": [1, 4, 5], "filter": 1, "thi": [1, 2, 3, 4, 5, 6, 7], "function": [1, 3, 5], "current_target": 1, "base_target": [1, 3, 5], "basetarget": [1, 4, 5, 7], "connect": [1, 2, 5], "handl": [1, 3, 4, 6], "global": [1, 5], "onli": [1, 2, 3, 5, 6], "bind": 1, "cdp": [1, 3, 5], "protocol": [1, 3, 5], "mv3_extens": [1, 2], "background": 1, "script": [1, 3, 5, 6, 7], "load": [1, 2, 3, 4, 5], "extens": [1, 2, 3, 4, 5], "manifest": 1, "version": 1, "3": [1, 6], "ensure_extensions_incognito_allow": 1, "ensur": [1, 6], "instal": 1, "allow": [1, 3, 5], "gener": [1, 6], "decid": 1, "whether": [1, 2, 3, 4, 5, 6], "split": 1, "span": 1, "not_allow": 1, "configur": 1, "For": [1, 3, 5], "chang": [1, 4, 6], "behaviour": [1, 3], "you": [1, 2, 3, 4, 5, 6], "ll": 1, "have": 1, "modifi": 1, "json": [1, 2, 3, 4, 5, 6], "file": [1, 2, 3, 5, 6], "compress": [1, 2, 4], "directori": [1, 2, 3, 5], "see": [1, 3, 4, 5, 6, 7], "develop": 1, "com": [1, 2, 7], "doc": 1, "refer": [1, 3, 4, 5, 6], "base_context": 1, "which": [1, 2, 3, 4, 5], "isn": [1, 6], "downloads_dir": [1, 2, 3, 5], "download": [1, 2, 3, 5], "set_download_behaviour": [1, 3], "path": [1, 2, 3, 5, 6], "deni": [1, 3], "allowandnam": [1, 3], "instead": [1, 2, 3, 6], "can": [1, 2, 3, 5, 6], "caus": [1, 3, 4], "some": [1, 3, 4, 7], "bug": [1, 3], "current_context": 1, "switch": [1, 3], "target_id": [1, 3, 5], "2": [1, 3, 4, 5, 6, 7], "targetid": [1, 5], "advanc": 1, "usag": [1, 3, 5, 6], "float": [1, 3, 5, 6], "second": [1, 3, 5, 6], "": [1, 6], "track": 1, "alreadi": [1, 4], "get_target_for_ifram": 1, "ifram": [1, 3, 5, 6], "specif": [1, 3, 5], "oopif": 1, "site": [1, 4], "isol": [1, 5], "solut": 1, "look": 1, "webel": [1, 3, 5, 7], "content_docu": [1, 6], "get_targets_for_ifram": 1, "more": 1, "inform": 1, "wait_download": [1, 3, 5], "wait": [1, 3, 5, 6, 7], "someth": [1, 3, 5], "like": [1, 3, 5, 6], "frameid": [1, 3, 4, 5], "2d543b5e8b14945b280c537a4882a695": [1, 3, 5], "guid": [1, 3, 5], "c91df4d5": [1, 3, 5], "9b45": [1, 3, 5], "4962": [1, 3, 5], "84df": [1, 3, 5], "3749bd3f926d": [1, 3, 5], "http": [1, 2, 3, 4, 5, 6, 7], "www": [1, 2, 3, 5], "w3": [1, 2, 3, 4, 5], "org": [1, 2, 3, 4, 5, 6], "wai": [1, 2, 3, 5], "er": [1, 2, 3, 5], "test": [1, 2, 3, 5, 6], "xhtml": [1, 2, 3, 5], "testfil": [1, 2, 3, 5], "resourc": [1, 2, 3, 4, 5], "pdf": [1, 2, 3, 5], "dummi": [1, 2, 3, 5], "suggestedfilenam": [1, 3, 5], "specifi": [1, 3, 4, 5], "guid_fil": [1, 2, 3, 5], "d": [1, 3, 5], "system": [1, 3, 5], "appdata": [1, 3, 5], "pycharm": [1, 3, 5], "scratch": [1, 3, 5], "time": [1, 3, 4, 5, 6], "from": [1, 2, 3, 4, 5, 6, 7], "yet": [1, 2, 3, 5, 6, 7], "referr": [1, 3, 4, 5], "wait_load": [1, 3, 4, 5, 7], "web": [1, 3, 5, 6], "bool": [1, 2, 3, 4, 5, 6], "webpag": [1, 3, 5, 6], "maximum": [1, 3, 5, 6], "same": [1, 3, 4, 5], "initi": [1, 3, 4, 5], "titl": [1, 3, 5, 7], "current_point": 1, "pointer": 1, "execute_raw_script": [1, 3, 5, 6], "arg": [1, 3, 5, 6], "await_r": [1, 3, 5], "serial": [1, 3, 5, 6], "deep": [1, 5], "max_depth": [1, 3, 5, 6], "execution_context_id": [1, 3, 5, 6], "unique_context": [1, 3, 5, 6], "execut": [1, 2, 3, 5, 6], "javascript": [1, 3, 5, 6], "globalthi": [1, 3, 5, 6], "argument": [1, 2, 3, 5, 6], "document": [1, 3, 5, 6], "obj": [1, 3, 5, 6], "here": [1, 3, 5, 6], "string": [1, 3, 4, 5], "pass": [1, 2, 4, 5], "either": [1, 2, 5], "serializ": [1, 5], "remoteobject": [1, 5], "await": [1, 2, 3, 4, 5, 6, 7], "valu": [1, 2, 3, 4, 5, 6], "idonli": [1, 5], "one": [1, 3, 4, 5, 6], "int": [1, 2, 3, 4, 5, 6], "depth": [1, 5, 6], "object": [1, 3, 4, 5], "complet": [1, 5, 7], "run": [1, 4, 5, 7], "exclus": [1, 5], "runtim": [1, 5, 6], "callfunctionon": [1, 5], "execute_script": [1, 3, 5, 6], "synchron": [1, 3, 5, 6], "descript": [1, 3, 5, 6], "execute_async_script": [1, 3, 5, 6], "asynchron": [1, 3, 5, 6], "i": [1, 2, 3, 4, 5, 6, 7], "recommend": [1, 3, 6], "except": [1, 3, 6], "correctli": [1, 3, 6], "eval_async": [1, 3, 5, 6], "resolv": [1, 3, 5, 6], "length": [1, 3, 5, 6], "1": [1, 3, 4, 5, 6], "re": [1, 3, 5, 6], "fetch": [1, 3, 4, 5, 6], "httpbin": [1, 3, 4, 5, 6], "mind": [1, 3, 4, 5, 6], "cor": [1, 3, 5, 6], "current_url": [1, 3, 5], "page_sourc": [1, 3, 5], "html": [1, 3, 5, 6, 7], "close": [1, 3, 5, 6], "focu": [1, 3, 6], "focus": [1, 6], "quit": [1, 3], "clean_dir": 1, "clean": [1, 2], "out": 1, "user": [1, 2, 4, 6], "data": [1, 2, 3, 4, 5, 6], "dir": [1, 2], "current_target_info": 1, "current_window_handl": [1, 3], "deprec": [1, 2], "remov": [1, 2, 3, 5, 6], "current_window_id": 1, "windowid": 1, "belong": 1, "window_handl": [1, 3], "aren": [1, 7], "order": [1, 4], "posit": [1, 3], "do": [1, 6], "reli": 1, "index": 1, "iter": [1, 3, 4, 5], "them": 1, "new_window": [1, 3], "type_hint": [1, 3], "activ": [1, 3], "should": [1, 2, 3, 4, 5], "explicitli": 1, "new_target": 1, "set_window_st": 1, "state": 1, "normal": 1, "minim": [1, 3], "maxim": [1, 3, 5], "fullscreen": 1, "normalize_window": 1, "size": [1, 6], "maximize_window": [1, 3], "fullscreen_window": [1, 3], "enter": [1, 4], "minimize_window": [1, 3], "rtype": [1, 2], "throttl": 1, "print_pag": [1, 3, 5], "print": [1, 2, 4, 5, 7], "switch_to": [1, 3, 7], "switchto": [1, 3, 7], "back": [1, 3, 5], "goe": [1, 3, 5], "step": [1, 3, 5], "backward": [1, 3, 5], "histori": [1, 3, 5], "ha": [1, 2, 4, 5, 6], "forward": [1, 3, 5], "refresh": [1, 3, 5], "get_cooki": [1, 3, 5], "cooki": [1, 3, 5], "singl": [1, 2, 3, 4, 5], "found": [1, 3, 5], "delete_cooki": [1, 3, 5], "domain": [1, 3, 5], "delet": [1, 3, 5], "given": [1, 3, 5, 6], "delete_all_cooki": [1, 3, 5], "add_cooki": [1, 3, 5], "cookie_dict": [1, 3, 5], "add": [1, 2, 3, 5], "network": [1, 3, 4, 5], "cookieparam": [1, 5], "static": [1, 4], "sleep": [1, 7], "time_to_wait": 1, "sync": 1, "find_el": [1, 3, 5, 6, 7], "find": [1, 3, 5, 6], "element": [1, 3, 5, 6, 7], "locat": [1, 3, 5, 6, 7], "By": [1, 2, 3, 4, 5, 6, 7], "actual": [1, 3, 5, 6], "queri": [1, 3, 5, 6], "how": [1, 3, 5, 6], "long": [1, 3, 4, 5, 6], "exist": [1, 3, 5, 6, 7], "multipl": [1, 2, 3, 5, 6], "search_el": [1, 3, 5], "similarli": 1, "ctrl": 1, "f": 1, "devtool": [1, 2, 3, 5], "consol": 1, "plain": [1, 3, 5], "text": [1, 3, 4, 5, 6, 7], "get_screenshot_as_fil": [1, 3, 5], "filenam": [1, 3, 5, 6], "save": [1, 2, 3, 5, 6], "screenshot": [1, 3, 5, 6], "png": [1, 3, 5, 6], "imag": [1, 3, 5, 6], "wish": [1, 3, 5], "your": [1, 3, 4, 5], "end": [1, 3, 5], "save_screenshot": [1, 3, 5], "func": 1, "get_screenshot_as_png": [1, 3, 5], "binari": [1, 2, 3, 4, 5, 6], "byte": [1, 3, 4, 5, 6], "get_screenshot_as_base64": [1, 3, 5], "base64": [1, 3, 4, 5, 6], "encod": [1, 3, 4, 5], "embed": [1, 3, 5], "set_window_s": [1, 3], "width": [1, 3], "height": [1, 3], "unless": [1, 2], "windowhandl": [1, 3], "pixel": [1, 3], "get_window_s": [1, 3], "code": [1, 4], "block": [1, 4], "1280": 1, "720": 1, "set_window_posit": [1, 3], "x": [1, 3], "y": [1, 3], "coordin": [1, 3, 6], "get_window_posit": [1, 3], "0": [1, 3, 4, 6, 7], "get_window_rect": [1, 3], "windowst": 1, "set_window_rect": [1, 3], "get_network_condit": [1, 3, 5], "emul": [1, 3, 5], "latenc": [1, 3, 5], "4": [1, 3, 5], "download_throughput": [1, 3, 5], "upload_throughput": [1, 3, 5], "offlin": [1, 3, 5], "set_network_condit": [1, 3, 5], "connection_typ": [1, 3, 5], "addit": [1, 3, 5], "m": [1, 3, 5, 7], "throughput": [1, 3, 5], "500": [1, 3, 5], "1024": [1, 3, 5], "exampl": [1, 2, 3, 5, 7], "cellular2g": 1, "cellular3g": 1, "cellular4g": 1, "bluetooth": 1, "ethernet": 1, "wifi": [1, 3, 5], "wimax": 1, "both": [1, 3, 5], "upload": [1, 3, 5], "delete_network_condit": [1, 3, 5], "reset": [1, 3, 5], "set_permiss": [1, 3], "applic": [1, 3, 4], "permiss": [1, 3], "item": [1, 3], "prompt": 1, "appli": [1, 2], "clipboard": [1, 3], "read": [1, 3], "set_proxi": 1, "proxy_config": 1, "dynam": [1, 5], "mode": 1, "fixed_serv": 1, "rule": 1, "proxyforhttp": 1, "scheme": [1, 4], "host": [1, 2, 5], "port": [1, 2], "proxyforftp": 1, "fallbackproxi": 1, "bypasslist": 1, "local": 1, "authentif": [1, 4], "set_auth": 1, "set_single_proxi": 1, "bypass_list": 1, "user1": [1, 2], "passwrd1": [1, 2], "5001": [1, 2], "when": [1, 2, 3, 4, 5], "been": [1, 2, 4], "driverless": [1, 2], "socks5": [1, 2], "authent": [1, 2, 4], "1309413": [1, 2], "clear_proxi": 1, "clear": [1, 6], "usernam": [1, 4], "password": [1, 4], "host_with_port": 1, "format": [1, 2, 6], "clear_auth": 1, "auth": [1, 4], "wait_for_cdp": [1, 3, 5, 7], "event": [1, 3, 4, 5, 6], "even": [], "windowopen": [], "add_cdp_listen": [1, 3, 5], "callback": [1, 3, 4, 5], "listen": [1, 3, 5, 6], "callabl": [4, 5], "accept": [2, 5, 6, 7], "remove_cdp_listen": [1, 3, 5], "execute_cdp_cmd": [1, 3, 5], "cmd": [1, 3, 5], "cmd_arg": [1, 3, 5], "10": [1, 3, 5, 7], "command": [1, 3, 5, 6], "result": 5, "follow": 5, "link": [4, 5], "chromedevtool": 5, "github": [5, 7], "io": 5, "A": [1, 3, 5, 6], "empti": 5, "getresponsebodi": 5, "requestid": [4, 5], "base64encod": 5, "bodi": [4, 5, 7], "respons": [4, 5], "get_sink": [1, 3, 5], "sink": [1, 3, 5], "avail": [1, 3, 5], "cast": [1, 3, 5], "get_issue_messag": [1, 3, 5], "error": [1, 3, 4, 5], "messag": [1, 3, 5], "issu": [1, 3, 5], "session": [1, 2, 3, 5], "set_sink_to_us": [1, 3, 5], "sink_nam": [1, 3, 5], "its": [1, 3, 5], "receiv": [1, 3, 5], "start_desktop_mirror": [1, 3, 5], "desktop": [1, 3, 5], "mirror": [1, 3, 5], "start_tab_mirror": [1, 3, 5], "stop_cast": [1, 3, 5], "stop": [1, 3, 5], "loop": [1, 3, 5, 6], "alert": [1, 3, 7], "get_alert": 1, "5": [1, 2, 3, 5, 6, 7], "default_cont": [1, 3], "frame_refer": 1, "integ": 1, "repres": [1, 5, 6], "frame_nam": [1, 3], "union": [1, 2, 4, 6], "what": 1, "kind": 1, "parent_fram": [1, 3], "parent": [1, 6], "If": [1, 3, 6], "top": [1, 3], "level": [1, 3, 6], "brows": [1, 3], "remain": 1, "unchang": 1, "window_id": 1, "window_nam": 1, "main": [1, 3, 4, 7], "chrome": [2, 3, 4, 5, 6, 7], "add_argu": 2, "launch": 2, "pref": 2, "prefer": 2, "update_pref": 2, "updat": 2, "dot": 2, "user_data_dir": 2, "temporarili": 2, "temp": 2, "_dir": 2, "o": 2, "getcwd": 2, "isdir": 2, "mkdir": 2, "webdriv": [2, 3, 4, 7], "plugin": 2, "always_open_pdf_extern": 2, "download_data": 2, "absolut": [2, 6], "headless": 2, "startup_url": 2, "binary_loc": 2, "env": 2, "subprocess": 2, "popen": 2, "environ": 2, "add_extens": 2, "zip": 2, "crx": 2, "etc": 2, "extract": 2, "debugger_address": 2, "address": 2, "remot": 2, "instanc": [2, 5, 6], "make": [2, 3, 4, 5], "well": [2, 3], "single_proxi": 2, "auto_clean_dir": 2, "automat": [2, 3, 4, 6], "enable_mobil": 2, "android_packag": 2, "android": 2, "android_act": 2, "device_seri": 2, "enabl": [2, 6], "mobil": 2, "packag": 2, "Not": [1, 2, 3, 6], "implement": [2, 6], "accept_insecure_cert": 2, "insecur": 2, "certif": 2, "notimpl": [2, 6], "ignore_local_proxy_environment_vari": 2, "call": [2, 6], "http_proxi": 2, "https_proxi": 2, "being": [2, 6], "pick": 2, "up": 2, "add_experimental_opt": 2, "experiment": 2, "method": [2, 3, 4, 6, 7], "is_incognito": 3, "1048576": [3, 5], "drive": [3, 5], "chromedriv": [3, 5], "start_sess": 3, "desir": 3, "capabl": 3, "behavior": [3, 4], "click": [3, 6, 7], "const": 3, "accord": 3, "obj_id": [3, 6], "docs_sourc": 3, "start_monoton": 3, "everi": [3, 6], "associ": 3, "hint": 3, "select": [3, 6], "invok": 3, "manag": 3, "full": [3, 5], "screen": 3, "oper": [3, 6], "take": [3, 4, 5, 6], "best": [3, 5], "effort": [3, 5], "provid": [3, 4, 5], "contain": 3, "active_el": 3, "dictionari": [3, 5, 6], "correspond": [3, 4, 5], "visibl": [3, 5, 6], "my_cooki": 3, "scope": [3, 5], "requir": [3, 4], "kei": [3, 4], "secur": 3, "httponli": 3, "expiri": 3, "samesit": 3, "foo": [3, 5], "bar": 3, "strict": [3, 4], "search": [3, 5], "ioerror": [3, 5], "els": [3, 4, 5, 6], "resizeto": 3, "800": 3, "600": 3, "moveto": 3, "w3c": 3, "compat": 3, "100": 3, "200": [3, 4], "network_condit": [3, 5], "condit": [3, 5], "note": [3, 4, 5], "network_interceptor": 4, "requestdoneexcept": 4, "authalreadyhandledexcept": 4, "requestpattern": 4, "enumer": 4, "is_remot": 5, "start_socket": 5, "get_histori": 5, "getnavigationhistori": 5, "navigationhistori": 5, "set_sourc": [5, 6], "15": [5, 7], "outerhtml": [5, 6], "dom": [5, 6], "try": 5, "might": [4, 5, 6], "fail": [4, 5, 6], "reload": 5, "target_info": 5, "target_gett": 5, "info": 5, "itself": 5, "attach": [5, 6], "client": 5, "opener_id": 5, "open": 5, "can_access_open": 5, "opener_frame_id": 5, "browser_context_id": 5, "browsercontextid": 5, "subtyp": 5, "detail": [5, 6], "mai": 5, "portal": 5, "prerend": 5, "frame_id": [4, 6], "isolated_exec_id": 6, "node_id": 6, "backend_node_id": 6, "is_ifram": 6, "interest": 6, "interact": 6, "perform": 6, "through": 6, "interfac": 6, "fresh": 6, "check": 6, "still": [4, 6], "valid": 6, "essenti": 6, "determin": 6, "staleelementreferenceexcept": 6, "thrown": 6, "futur": 6, "remoteobjectid": 6, "document_url": 6, "idx": 6, "get_listen": 6, "domdebugg": 6, "geteventlisten": 6, "get_properti": 6, "elem": [6, 7], "tagnam": 6, "textcont": 6, "entri": 6, "highlight": 6, "disabl": 6, "fade": 6, "is_click": 6, "listener_depth": 6, "button": 6, "input": 6, "textarea": 6, "video": 6, "map": 6, "wise": 6, "mousedown": 6, "mouseup": 6, "visible_timeout": 6, "spread_a": 6, "spread_b": 6, "bias_a": 6, "bias_b": 6, "border": 6, "05": 6, "scroll_to": 6, "move_to": [6, 7], "ensure_click": 6, "abl": 6, "comput": 6, "box": 6, "model": 6, "spread": 6, "over": [1, 3, 4, 5, 6], "b": 6, "bia": 6, "minimum": 6, "toward": 6, "edg": 6, "rel": 6, "random": 6, "point": 6, "outsid": 6, "scroll": 6, "move": 6, "mous": 6, "clickabl": 6, "reliabl": 6, "equival": 6, "6": 6, "std": 6, "99": 6, "7": 6, "write": 6, "insert": 6, "send": 6, "set_fil": 6, "send_kei": 6, "mid_loc": 6, "probabl": 6, "middl": 6, "submit": 6, "form": 6, "detect": [4, 6], "It": 6, "possibl": [4, 6], "get_dom_attribut": 6, "attribut": 6, "unlik": 6, "get_attribut": 6, "declar": 6, "markup": 6, "retriev": 6, "text_length": 6, "target_el": 6, "is_select": 6, "checkbox": 6, "radio": 6, "is_en": 6, "shadow_root": 6, "shadowroot": 6, "doe": [4, 6], "shadow": 6, "is_displai": 6, "location_once_scrolled_into_view": 6, "rect": 6, "properli": [6, 7], "value_of_css_properti": 6, "property_nam": 6, "render": 6, "canva": 6, "box_model": 6, "boxmodel": 6, "aria_rol": 6, "aria": 6, "role": 6, "accessible_nam": 6, "screenshot_as_base64": 6, "screenshot_as_png": 6, "chromeopt": 7, "request": 7, "intercept": 7, "networkinterceptor": 4, "on_request": 4, "on_respons": 4, "on_auth": 4, "pattern": 4, "intercept_auth": 4, "bypass_service_work": 4, "__init__": 4, "interceptedrequest": 4, "onrequest": 4, "onrespons": 4, "onauth": 4, "bypass": 4, "servic": 4, "worker": 4, "patter": 4, "__aiter__": 4, "asynciter": [1, 3, 4, 5], "interceptedauth": 4, "ware": [], "virtual": 4, "zero": 4, "asyncio": [4, 7], "ensure_futur": 4, "where": 4, "warn": [], "import": [4, 7], "requeststag": 4, "def": [4, 7], "post": 4, "continue_request": 4, "basic": 4, "intercept_respons": 4, "anyrequest": 4, "interceptor": 4, "isinst": 4, "would": 4, "continue_auth": 4, "elif": 4, "stage": 4, "break": 4, "python": 7, "requestpaus": 4, "authrequir": 4, "get_cdp_event_it": [1, 3, 5], "socket": 5, "singlecdpsocket": 5, "framenavig": 5, "now": 5, "chromeinst": 5, "whole": 5, "usual": 5, "unfortun": 5, "downloads_dir_for_context": 5, "is_redirect": 4, "redirect": 4, "response_status_cod": 4, "301": 4, "302": 4, "303": 4, "307": 4, "308": 4, "resource_typ": 4, "resourcetyp": 4, "network_id": 4, "requestwillbes": 4, "had": 4, "fire": 4, "response_error_reason": 4, "reason": 4, "response_head": 4, "header": 4, "statu": 4, "response_status_text": 4, "redirected_id": 4, "bypass_brows": 4, "allow_redirect": 4, "proxy_auth": 4, "extern": 4, "tl": 4, "fingerprint": 4, "accordingli": 4, "technic": 4, "post_data": 4, "continu": 4, "fulfil": 4, "arrai": 4, "observ": 4, "overrid": 4, "continue_respons": 4, "binary_response_head": 4, "response_cod": 4, "response_phras": 4, "separ": 4, "seri": 4, "pair": 4, "treat": 4, "phrase": 4, "ok": 4, "resum": 4, "rais": 4, "cancel": 4, "done": 4, "fail_request": 4, "error_reason": 4, "abort": 4, "timedout": 4, "accessdeni": 4, "connectionclos": 4, "connectionreset": 4, "connectionrefus": 4, "connectionabort": 4, "connectionfail": 4, "namenotresolv": 4, "internetdisconnect": 4, "addressunreach": 4, "blockedbycli": 4, "blockedbyrespons": 4, "auth_challeng": 4, "authchalleng": 4, "cancelauth": 4, "providecredenti": 4, "challeng": 4, "digest": 4, "realm": 4, "pleas": 4, "fragment": 4, "url_frag": 4, "hash": 4, "present": 4, "has_post_data": 4, "postdata": 4, "omit": 4, "flag": 4, "too": 4, "post_data_entri": 4, "postdataentri": 4, "mixed_content_typ": 4, "blockabl": 4, "mix": 4, "content": 4, "initial_prior": 4, "verylow": 4, "low": 4, "medium": 4, "high": 4, "veryhigh": 4, "prioriti": 4, "sent": 4, "referrer_polici": 4, "unsaf": 4, "downgrad": 4, "polici": 4, "defin": 4, "tr": 4, "is_link_preload": 4, "via": 4, "preload": 4, "trust_token_param": 4, "trusttokenparam": 4, "is_same_sit": 4, "consid": 4, "urlpattern": 4, "anyrespons": 4, "url_pattern": 4, "request_stag": 4, "nowsecur": 7, "nl": 7, "relax": 7, "domcontenteventfir": 7, "div": 7, "p": 7, "kaliiiiiiiiii": 7, "selenium": 7, "pip": 7, "upgrad": 7}, "objects": {"selenium_driverless.scripts.network_interceptor": [[4, 0, 1, "", "AuthAlreadyHandledException"], [4, 0, 1, "", "AuthChallenge"], [4, 0, 1, "", "InterceptedAuth"], [4, 0, 1, "", "InterceptedRequest"], [4, 0, 1, "", "NetworkInterceptor"], [4, 0, 1, "", "Request"], [4, 0, 1, "", "RequestDoneException"], [4, 0, 1, "", "RequestPattern"], [4, 0, 1, "", "RequestStages"]], "selenium_driverless.scripts.network_interceptor.AuthAlreadyHandledException": [[4, 1, 1, "", "request"]], "selenium_driverless.scripts.network_interceptor.AuthChallenge": [[4, 1, 1, "", "origin"], [4, 1, 1, "", "params"], [4, 1, 1, "", "realm"], [4, 1, 1, "", "scheme"], [4, 1, 1, "", "source"], [4, 1, 1, "", "target"]], "selenium_driverless.scripts.network_interceptor.InterceptedAuth": [[4, 1, 1, "", "auth_challenge"], [4, 2, 1, "", "cancel"], [4, 2, 1, "", "continue_auth"], [4, 1, 1, "", "frame_id"], [4, 1, 1, "", "id"], [4, 1, 1, "", "params"], [4, 1, 1, "", "request"], [4, 1, 1, "", "resource_type"], [4, 2, 1, "", "resume"], [4, 1, 1, "", "target"]], "selenium_driverless.scripts.network_interceptor.InterceptedRequest": [[4, 1, 1, "", "body"], [4, 2, 1, "", "bypass_browser"], [4, 2, 1, "", "continue_request"], [4, 2, 1, "", "continue_response"], [4, 2, 1, "", "fail_request"], [4, 1, 1, "", "frame_id"], [4, 2, 1, "", "fulfill"], [4, 1, 1, "", "id"], [4, 1, 1, "", "is_redirect"], [4, 1, 1, "", "network_id"], [4, 1, 1, "", "params"], [4, 1, 1, "", "redirected_id"], [4, 1, 1, "", "request"], [4, 1, 1, "", "resource_type"], [4, 1, 1, "", "response_error_reason"], [4, 1, 1, "", "response_headers"], [4, 1, 1, "", "response_status_code"], [4, 1, 1, "", "response_status_text"], [4, 2, 1, "", "resume"], [4, 1, 1, "", "stage"], [4, 1, 1, "", "target"]], "selenium_driverless.scripts.network_interceptor.NetworkInterceptor": [[4, 2, 1, "", "__aiter__"], [4, 2, 1, "", "__init__"], [4, 1, 1, "", "patterns"], [4, 1, 1, "", "target"]], "selenium_driverless.scripts.network_interceptor.Request": [[4, 1, 1, "", "has_post_data"], [4, 1, 1, "", "headers"], [4, 1, 1, "", "initial_priority"], [4, 1, 1, "", "is_link_preload"], [4, 1, 1, "", "is_same_site"], [4, 1, 1, "", "method"], [4, 1, 1, "", "mixed_content_type"], [4, 1, 1, "", "params"], [4, 1, 1, "", "post_data"], [4, 1, 1, "", "post_data_entries"], [4, 1, 1, "", "referrer_policy"], [4, 1, 1, "", "target"], [4, 1, 1, "", "trust_token_params"], [4, 1, 1, "", "url"], [4, 1, 1, "", "url_fragment"]], "selenium_driverless.scripts.network_interceptor.RequestDoneException": [[4, 1, 1, "", "request"]], "selenium_driverless.scripts.network_interceptor.RequestPattern": [[4, 3, 1, "", "AnyRequest"], [4, 3, 1, "", "AnyResponse"], [4, 2, 1, "", "new"]], "selenium_driverless.scripts.network_interceptor.RequestStages": [[4, 3, 1, "", "Request"], [4, 3, 1, "", "Response"]], "selenium_driverless.scripts.switch_to": [[1, 0, 1, "", "SwitchTo"]], "selenium_driverless.scripts.switch_to.SwitchTo": [[1, 1, 1, "", "alert"], [1, 2, 1, "", "default_content"], [1, 2, 1, "", "frame"], [1, 2, 1, "", "get_alert"], [1, 2, 1, "", "new_window"], [1, 2, 1, "", "parent_frame"], [1, 2, 1, "", "target"], [1, 2, 1, "", "window"]], "selenium_driverless.types.base_target": [[5, 0, 1, "", "BaseTarget"]], "selenium_driverless.types.base_target.BaseTarget": [[5, 2, 1, "", "add_cdp_listener"], [5, 2, 1, "", "downloads_dir_for_context"], [5, 2, 1, "", "execute_cdp_cmd"], [5, 2, 1, "", "get_cdp_event_iter"], [5, 2, 1, "", "remove_cdp_listener"], [5, 1, 1, "", "socket"], [5, 2, 1, "", "wait_for_cdp"]], "selenium_driverless.types.by": [[0, 0, 1, "", "By"]], "selenium_driverless.types.by.By": [[0, 3, 1, "", "CLASS_NAME"], [0, 3, 1, "", "CSS"], [0, 3, 1, "", "CSS_SELECTOR"], [0, 3, 1, "", "ID"], [0, 3, 1, "", "NAME"], [0, 3, 1, "", "TAG_NAME"], [0, 3, 1, "", "XPATH"]], "selenium_driverless.types.context": [[3, 0, 1, "", "Context"]], "selenium_driverless.types.context.Context": [[3, 2, 1, "", "add_cdp_listener"], [3, 2, 1, "", "add_cookie"], [3, 2, 1, "", "back"], [3, 2, 1, "", "close"], [3, 1, 1, "", "current_url"], [3, 1, 1, "", "current_window_handle"], [3, 2, 1, "", "delete_all_cookies"], [3, 2, 1, "", "delete_cookie"], [3, 2, 1, "", "delete_network_conditions"], [3, 1, 1, "", "downloads_dir"], [3, 2, 1, "", "eval_async"], [3, 2, 1, "", "execute_async_script"], [3, 2, 1, "", "execute_cdp_cmd"], [3, 2, 1, "", "execute_raw_script"], [3, 2, 1, "", "execute_script"], [3, 2, 1, "", "find_element"], [3, 2, 1, "", "find_elements"], [3, 2, 1, "", "forward"], [3, 2, 1, "", "fullscreen_window"], [3, 2, 1, "", "get"], [3, 2, 1, "", "get_cdp_event_iter"], [3, 2, 1, "", "get_cookie"], [3, 2, 1, "", "get_cookies"], [3, 2, 1, "", "get_issue_message"], [3, 2, 1, "", "get_network_conditions"], [3, 2, 1, "", "get_screenshot_as_base64"], [3, 2, 1, "", "get_screenshot_as_file"], [3, 2, 1, "", "get_screenshot_as_png"], [3, 2, 1, "", "get_sinks"], [3, 2, 1, "", "get_window_position"], [3, 2, 1, "", "get_window_rect"], [3, 2, 1, "", "get_window_size"], [3, 2, 1, "", "maximize_window"], [3, 2, 1, "", "minimize_window"], [3, 2, 1, "", "new_window"], [3, 1, 1, "", "page_source"], [3, 2, 1, "", "print_page"], [3, 2, 1, "", "quit"], [3, 2, 1, "", "refresh"], [3, 2, 1, "", "remove_cdp_listener"], [3, 2, 1, "", "save_screenshot"], [3, 2, 1, "", "search_elements"], [3, 2, 1, "", "set_download_behaviour"], [3, 2, 1, "", "set_network_conditions"], [3, 2, 1, "", "set_permissions"], [3, 2, 1, "", "set_sink_to_use"], [3, 2, 1, "", "set_window_position"], [3, 2, 1, "", "set_window_rect"], [3, 2, 1, "", "set_window_size"], [3, 2, 1, "", "start_desktop_mirroring"], [3, 2, 1, "", "start_session"], [3, 2, 1, "", "start_tab_mirroring"], [3, 2, 1, "", "stop_casting"], [3, 1, 1, "", "switch_to"], [3, 1, 1, "", "title"], [3, 2, 1, "", "wait_download"], [3, 2, 1, "", "wait_for_cdp"], [3, 1, 1, "", "window_handles"]], "selenium_driverless.types.options": [[2, 0, 1, "", "Options"]], "selenium_driverless.types.options.Options": [[2, 1, 1, "", "accept_insecure_certs"], [2, 2, 1, "", "add_argument"], [2, 2, 1, "", "add_arguments"], [2, 2, 1, "", "add_experimental_option"], [2, 2, 1, "", "add_extension"], [2, 1, 1, "", "arguments"], [2, 1, 1, "", "auto_clean_dirs"], [2, 1, 1, "", "binary_location"], [2, 1, 1, "", "debugger_address"], [2, 1, 1, "", "downloads_dir"], [2, 2, 1, "", "enable_mobile"], [2, 1, 1, "", "env"], [2, 1, 1, "", "headless"], [2, 2, 1, "", "ignore_local_proxy_environment_variables"], [2, 1, 1, "", "prefs"], [2, 1, 1, "", "single_proxy"], [2, 1, 1, "", "startup_url"], [2, 2, 1, "", "update_pref"], [2, 1, 1, "", "user_data_dir"]], "selenium_driverless.types.target": [[5, 0, 1, "", "Target"], [5, 0, 1, "", "TargetInfo"]], "selenium_driverless.types.target.Target": [[5, 2, 1, "", "add_cdp_listener"], [5, 2, 1, "", "add_cookie"], [5, 2, 1, "", "back"], [5, 2, 1, "", "close"], [5, 1, 1, "", "current_url"], [5, 2, 1, "", "delete_all_cookies"], [5, 2, 1, "", "delete_cookie"], [5, 2, 1, "", "delete_network_conditions"], [5, 2, 1, "", "eval_async"], [5, 2, 1, "", "execute_async_script"], [5, 2, 1, "", "execute_cdp_cmd"], [5, 2, 1, "", "execute_raw_script"], [5, 2, 1, "", "execute_script"], [5, 2, 1, "", "find_element"], [5, 2, 1, "", "find_elements"], [5, 2, 1, "", "forward"], [5, 2, 1, "", "get"], [5, 2, 1, "", "get_cdp_event_iter"], [5, 2, 1, "", "get_cookie"], [5, 2, 1, "", "get_cookies"], [5, 2, 1, "", "get_history"], [5, 2, 1, "", "get_issue_message"], [5, 2, 1, "", "get_network_conditions"], [5, 2, 1, "", "get_screenshot_as_base64"], [5, 2, 1, "", "get_screenshot_as_file"], [5, 2, 1, "", "get_screenshot_as_png"], [5, 2, 1, "", "get_sinks"], [5, 1, 1, "", "page_source"], [5, 2, 1, "", "print_page"], [5, 2, 1, "", "refresh"], [5, 2, 1, "", "remove_cdp_listener"], [5, 2, 1, "", "save_screenshot"], [5, 2, 1, "", "search_elements"], [5, 2, 1, "", "set_network_conditions"], [5, 2, 1, "", "set_sink_to_use"], [5, 2, 1, "", "set_source"], [5, 1, 1, "", "socket"], [5, 2, 1, "", "start_desktop_mirroring"], [5, 2, 1, "", "start_tab_mirroring"], [5, 2, 1, "", "stop_casting"], [5, 1, 1, "", "title"], [5, 2, 1, "", "wait_download"], [5, 2, 1, "", "wait_for_cdp"]], "selenium_driverless.types.target.TargetInfo": [[5, 1, 1, "", "Target"], [5, 1, 1, "", "attached"], [5, 1, 1, "", "browser_context_id"], [5, 1, 1, "", "can_access_opener"], [5, 1, 1, "", "id"], [5, 1, 1, "", "opener_frame_id"], [5, 1, 1, "", "opener_id"], [5, 1, 1, "", "subtype"]], "selenium_driverless.types.webelement": [[6, 0, 1, "", "WebElement"]], "selenium_driverless.types.webelement.WebElement": [[6, 1, 1, "", "accessible_name"], [6, 1, 1, "", "aria_role"], [6, 1, 1, "", "box_model"], [6, 2, 1, "", "clear"], [6, 2, 1, "", "click"], [6, 1, 1, "", "content_document"], [6, 1, 1, "", "document_url"], [6, 2, 1, "", "eval_async"], [6, 2, 1, "", "execute_async_script"], [6, 2, 1, "", "execute_script"], [6, 2, 1, "", "find_element"], [6, 2, 1, "", "find_elements"], [6, 2, 1, "", "focus"], [6, 2, 1, "", "get_attribute"], [6, 2, 1, "", "get_dom_attribute"], [6, 2, 1, "", "get_listeners"], [6, 2, 1, "", "get_property"], [6, 2, 1, "", "highlight"], [6, 2, 1, "", "is_clickable"], [6, 2, 1, "", "is_displayed"], [6, 2, 1, "", "is_enabled"], [6, 2, 1, "", "is_selected"], [6, 1, 1, "", "location"], [6, 1, 1, "", "location_once_scrolled_into_view"], [6, 2, 1, "", "mid_location"], [6, 1, 1, "", "obj_id"], [6, 1, 1, "", "parent"], [6, 1, 1, "", "rect"], [6, 2, 1, "", "remove"], [6, 2, 1, "", "screenshot"], [6, 1, 1, "", "screenshot_as_base64"], [6, 1, 1, "", "screenshot_as_png"], [6, 2, 1, "", "scroll_to"], [6, 2, 1, "", "send_keys"], [6, 2, 1, "", "set_file"], [6, 2, 1, "", "set_files"], [6, 2, 1, "", "set_source"], [6, 1, 1, "", "shadow_root"], [6, 1, 1, "", "size"], [6, 1, 1, "", "source"], [6, 2, 1, "", "submit"], [6, 1, 1, "", "tag_name"], [6, 1, 1, "", "text"], [6, 1, 1, "", "value"], [6, 2, 1, "", "value_of_css_property"], [6, 2, 1, "", "write"]], "selenium_driverless.webdriver": [[1, 0, 1, "", "Chrome"]], "selenium_driverless.webdriver.Chrome": [[1, 2, 1, "", "add_cdp_listener"], [1, 2, 1, "", "add_cookie"], [1, 2, 1, "", "back"], [1, 1, 1, "", "base_context"], [1, 1, 1, "", "base_target"], [1, 2, 1, "", "clear_auth"], [1, 2, 1, "", "clear_proxy"], [1, 2, 1, "", "close"], [1, 1, 1, "", "contexts"], [1, 1, 1, "", "current_context"], [1, 1, 1, "", "current_pointer"], [1, 1, 1, "", "current_target"], [1, 1, 1, "", "current_target_info"], [1, 1, 1, "", "current_url"], [1, 1, 1, "", "current_window_handle"], [1, 1, 1, "", "current_window_id"], [1, 2, 1, "", "delete_all_cookies"], [1, 2, 1, "", "delete_cookie"], [1, 2, 1, "", "delete_network_conditions"], [1, 1, 1, "", "downloads_dir"], [1, 2, 1, "", "ensure_extensions_incognito_allowed"], [1, 2, 1, "", "eval_async"], [1, 2, 1, "", "execute_async_script"], [1, 2, 1, "", "execute_cdp_cmd"], [1, 2, 1, "", "execute_raw_script"], [1, 2, 1, "", "execute_script"], [1, 2, 1, "", "find_element"], [1, 2, 1, "", "find_elements"], [1, 2, 1, "", "focus"], [1, 2, 1, "", "forward"], [1, 1, 1, "", "frame_tree"], [1, 2, 1, "", "fullscreen_window"], [1, 2, 1, "", "get"], [1, 2, 1, "", "get_cdp_event_iter"], [1, 2, 1, "", "get_cookie"], [1, 2, 1, "", "get_cookies"], [1, 2, 1, "", "get_issue_message"], [1, 2, 1, "", "get_network_conditions"], [1, 2, 1, "", "get_screenshot_as_base64"], [1, 2, 1, "", "get_screenshot_as_file"], [1, 2, 1, "", "get_screenshot_as_png"], [1, 2, 1, "", "get_sinks"], [1, 2, 1, "", "get_target"], [1, 2, 1, "", "get_target_for_iframe"], [1, 2, 1, "", "get_targets"], [1, 2, 1, "", "get_targets_for_iframes"], [1, 2, 1, "", "get_window_position"], [1, 2, 1, "", "get_window_rect"], [1, 2, 1, "", "get_window_size"], [1, 2, 1, "", "maximize_window"], [1, 2, 1, "", "minimize_window"], [1, 1, 1, "", "mv3_extension"], [1, 2, 1, "", "new_context"], [1, 2, 1, "", "new_window"], [1, 2, 1, "", "normalize_window"], [1, 1, 1, "", "page_source"], [1, 2, 1, "", "print_page"], [1, 2, 1, "", "quit"], [1, 2, 1, "", "refresh"], [1, 2, 1, "", "remove_cdp_listener"], [1, 2, 1, "", "save_screenshot"], [1, 2, 1, "", "search_elements"], [1, 2, 1, "", "set_auth"], [1, 2, 1, "", "set_download_behaviour"], [1, 2, 1, "", "set_network_conditions"], [1, 2, 1, "", "set_permissions"], [1, 2, 1, "", "set_proxy"], [1, 2, 1, "", "set_single_proxy"], [1, 2, 1, "", "set_sink_to_use"], [1, 2, 1, "", "set_window_position"], [1, 2, 1, "", "set_window_rect"], [1, 2, 1, "", "set_window_size"], [1, 2, 1, "", "set_window_state"], [1, 2, 1, "", "sleep"], [1, 2, 1, "", "start_desktop_mirroring"], [1, 2, 1, "", "start_tab_mirroring"], [1, 2, 1, "", "stop_casting"], [1, 1, 1, "", "switch_to"], [1, 1, 1, "", "targets"], [1, 1, 1, "", "title"], [1, 2, 1, "", "wait_download"], [1, 2, 1, "", "wait_for_cdp"], [1, 1, 1, "", "window_handles"]]}, "objtypes": {"0": "py:class", "1": "py:property", "2": "py:method", "3": "py:attribute"}, "objnames": {"0": ["py", "class", "Python class"], "1": ["py", "property", "Python property"], "2": ["py", "method", "Python method"], "3": ["py", "attribute", "Python attribute"]}, "titleterms": {"By": 0, "element": 0, "locat": 0, "webdriv": 1, "chrome": 1, "chromeopt": 2, "context": 3, "request": 4, "intercept": 4, "target": 5, "webel": 6, "document": 7, "selenium": [], "driverless": 7, "indic": [], "tabl": [], "api": [4, 7], "exampl": 4, "script": 4, "usag": 7, "instal": 7, "sourc": 7}, "envversion": {"sphinx.domains.c": 3, "sphinx.domains.changeset": 1, "sphinx.domains.citation": 1, "sphinx.domains.cpp": 9, "sphinx.domains.index": 1, "sphinx.domains.javascript": 3, "sphinx.domains.math": 2, "sphinx.domains.python": 4, "sphinx.domains.rst": 2, "sphinx.domains.std": 2, "sphinx.ext.viewcode": 1, "sphinx": 60}, "alltitles": {"By Element Locator": [[0, "by-element-locator"]], "ChromeOptions": [[2, "chromeoptions"]], "WebElement": [[6, "webelement"]], "webdriver.Chrome": [[1, "webdriver-chrome"]], "Context": [[3, "context"]], "Target": [[5, "target"]], "API": [[4, "api"], [7, "api"]], "Request-Interception": [[4, "request-interception"]], "Example Script": [[4, "example-script"]], "Documentation of Driverless": [[7, "documentation-of-driverless"]], "Installation": [[7, "installation"]], "Usage": [[7, "usage"]], "Source": [[7, "source"]]}, "indexentries": {}})
+Search.setIndex({"docnames": ["api/By", "api/Chrome", "api/ChromeOptions", "api/Context", "api/RequestInterception", "api/Target", "api/WebELement", "index"], "filenames": ["api/By.rst", "api/Chrome.rst", "api/ChromeOptions.rst", "api/Context.rst", "api/RequestInterception.rst", "api/Target.rst", "api/WebELement.rst", "index.rst"], "titles": ["By Element Locator", "webdriver.Chrome", "ChromeOptions", "Context", "Request-Interception", "Target", "WebElement", "Documentation of Driverless"], "terms": {"class": [0, 1, 2, 3, 4, 5, 6], "selenium_driverless": [0, 1, 2, 3, 4, 5, 6, 7], "type": [0, 1, 2, 3, 4, 5, 6, 7], "sourc": [0, 1, 2, 3, 4, 5, 6], "set": [0, 1, 2, 3, 4, 5, 6], "support": [0, 1, 2, 3, 5, 6], "strategi": 0, "id": [0, 1, 4, 5, 6], "name": [0, 1, 2, 3, 4, 5, 6], "xpath": [0, 3, 5, 7], "tag_nam": [0, 1, 3, 6], "tag": 0, "class_nam": [0, 6], "css_selector": 0, "css": 0, "selector": [0, 3, 5], "alia": [0, 1, 6], "option": [1, 2, 3, 4, 5, 6, 7], "none": [1, 2, 3, 4, 5, 6], "timeout": [1, 2, 3, 4, 5, 6, 7], "30": [1, 3, 4, 5, 6], "debug": 1, "fals": [1, 2, 3, 4, 5, 6], "max_ws_siz": [1, 3, 4, 5], "134217728": 1, "control": 1, "chromium": [1, 2, 3, 5], "base": [1, 3, 5], "browser": [1, 2, 3, 4, 5], "without": [1, 3, 4, 5], "ani": [1, 3, 4, 5, 6], "driver": [1, 2, 3, 4, 5, 7], "properti": [1, 2, 3, 4, 5, 6], "frame_tre": 1, "dict": [1, 2, 3, 4, 5, 6], "async": [1, 2, 3, 4, 5, 6, 7], "all": [1, 2, 3, 5, 6, 7], "nest": [1, 6], "frame": [1, 3, 4, 5], "within": [1, 3, 5], "current": [1, 3, 5, 6], "target": [1, 3, 4, 6, 7], "str": [1, 2, 3, 4, 5, 6], "targetinfo": [1, 3, 5, 7], "context": [1, 5, 7], "incognito": [1, 5], "new_context": 1, "proxy_bypass_list": 1, "proxy_serv": 1, "true": [1, 2, 3, 4, 5, 6, 7], "universal_access_origin": 1, "url": [1, 2, 3, 4, 5, 6], "about": [1, 2], "blank": [1, 2], "creat": [1, 2, 3, 4], "new": [1, 3, 4], "paramet": [1, 2, 3, 4, 5, 6], "first": [1, 2], "tab": [1, 2, 3, 5], "start": [1, 2, 3, 4, 5], "default": [1, 2, 3, 4, 5], "list": [1, 2, 3, 4, 5, 6], "an": [1, 2, 3, 4, 5, 6], "origin": [1, 3, 4, 5], "grant": 1, "unlimit": 1, "cross": [1, 4], "access": [1, 5], "part": 1, "other": [1, 3, 4], "than": 1, "those": [1, 5], "constitut": 1, "ar": [1, 5], "ignor": [1, 2], "return": [1, 2, 3, 4, 5, 6], "The": [1, 2, 3, 4, 5, 6], "proxi": [1, 2, 4], "doesn": [1, 2, 3, 4, 6], "t": [1, 2, 3, 4, 6, 7], "work": 1, "window": [1, 3, 5, 6], "due": [1, 2], "crbug": [1, 2], "1310057": 1, "server": [1, 4], "us": [1, 2, 3, 4, 5, 6], "get_target": 1, "_type": 1, "context_id": [1, 3, 5, 6], "self": 1, "get": [1, 2, 3, 4, 5, 6, 7], "liter": [1, 3, 4, 5, 6], "page": [1, 3, 4, 5, 6, 7], "background_pag": 1, "service_work": 1, "param": [1, 4, 5], "filter": 1, "thi": [1, 2, 3, 4, 5, 6, 7], "function": [1, 3, 5], "current_target": 1, "base_target": [1, 3, 5], "basetarget": [1, 4, 5, 7], "connect": [1, 2, 5], "handl": [1, 3, 4, 6], "global": [1, 5], "onli": [1, 2, 3, 5, 6], "bind": 1, "cdp": [1, 3, 5], "protocol": [1, 3, 5], "mv3_extens": [1, 2], "background": 1, "script": [1, 3, 5, 6, 7], "load": [1, 2, 3, 4, 5], "extens": [1, 2, 3, 4, 5], "manifest": 1, "version": 1, "3": [1, 6], "ensure_extensions_incognito_allow": 1, "ensur": [1, 6], "instal": 1, "allow": [1, 3, 5], "gener": [1, 6], "decid": 1, "whether": [1, 2, 3, 4, 5, 6], "split": 1, "span": 1, "not_allow": 1, "configur": 1, "For": [1, 3, 5], "chang": [1, 4, 6], "behaviour": [1, 3], "you": [1, 2, 3, 4, 5, 6], "ll": 1, "have": 1, "modifi": 1, "json": [1, 2, 3, 4, 5, 6], "file": [1, 2, 3, 5, 6], "compress": [1, 2, 4], "directori": [1, 2, 3, 5], "see": [1, 3, 4, 5, 6, 7], "develop": [1, 5], "com": [1, 2, 7], "doc": [1, 5], "refer": [1, 3, 4, 5, 6], "base_context": 1, "which": [1, 2, 3, 4, 5], "isn": [1, 6], "downloads_dir": [1, 2, 3, 5], "download": [1, 2, 3, 5], "set_download_behaviour": [1, 3], "path": [1, 2, 3, 5, 6], "deni": [1, 3], "allowandnam": [1, 3], "instead": [1, 2, 3, 6], "can": [1, 2, 3, 5, 6], "caus": [1, 3, 4], "some": [1, 3, 4, 7], "bug": [1, 3], "current_context": 1, "switch": [1, 3], "target_id": [1, 3, 5], "2": [1, 3, 4, 5, 6, 7], "targetid": [1, 5], "advanc": 1, "usag": [1, 3, 5, 6], "float": [1, 3, 5, 6], "second": [1, 3, 5, 6], "": [1, 6], "track": 1, "alreadi": [1, 4], "get_target_for_ifram": 1, "ifram": [1, 3, 5, 6], "specif": [1, 3, 5], "oopif": 1, "site": [1, 4], "isol": [1, 5], "solut": 1, "look": 1, "webel": [1, 3, 5, 7], "content_docu": [1, 6], "get_targets_for_ifram": 1, "more": 1, "inform": 1, "wait_download": [1, 3, 5], "wait": [1, 3, 5, 6, 7], "someth": [1, 3, 5], "like": [1, 3, 5, 6], "frameid": [1, 3, 4, 5], "2d543b5e8b14945b280c537a4882a695": [1, 3, 5], "guid": [1, 3, 5], "c91df4d5": [1, 3, 5], "9b45": [1, 3, 5], "4962": [1, 3, 5], "84df": [1, 3, 5], "3749bd3f926d": [1, 3, 5], "http": [1, 2, 3, 4, 5, 6, 7], "www": [1, 2, 3, 5], "w3": [1, 2, 3, 4, 5], "org": [1, 2, 3, 4, 5, 6], "wai": [1, 2, 3, 5], "er": [1, 2, 3, 5], "test": [1, 2, 3, 5, 6], "xhtml": [1, 2, 3, 5], "testfil": [1, 2, 3, 5], "resourc": [1, 2, 3, 4, 5], "pdf": [1, 2, 3, 5], "dummi": [1, 2, 3, 5], "suggestedfilenam": [1, 3, 5], "specifi": [1, 3, 4, 5], "guid_fil": [1, 2, 3, 5], "d": [1, 3, 5], "system": [1, 3, 5], "appdata": [1, 3, 5], "pycharm": [1, 3, 5], "scratch": [1, 3, 5], "time": [1, 3, 4, 5, 6], "from": [1, 2, 3, 4, 5, 6, 7], "yet": [1, 2, 3, 5, 6, 7], "referr": [1, 3, 4, 5], "wait_load": [1, 3, 4, 5, 7], "web": [1, 3, 5, 6], "bool": [1, 2, 3, 4, 5, 6], "webpag": [1, 3, 5, 6], "maximum": [1, 3, 5, 6], "same": [1, 3, 4, 5], "initi": [1, 3, 4, 5], "titl": [1, 3, 5, 7], "current_point": 1, "pointer": 1, "execute_raw_script": [1, 3, 5, 6], "arg": [1, 3, 5, 6], "await_r": [1, 3, 5], "serial": [1, 3, 5, 6], "deep": [1, 5], "max_depth": [1, 3, 5, 6], "execution_context_id": [1, 3, 5, 6], "unique_context": [1, 3, 5, 6], "execut": [1, 2, 3, 5, 6], "javascript": [1, 3, 5, 6], "globalthi": [1, 3, 5, 6], "argument": [1, 2, 3, 5, 6], "document": [1, 3, 5, 6], "obj": [1, 3, 5, 6], "here": [1, 3, 5, 6], "string": [1, 3, 4, 5], "pass": [1, 2, 4, 5], "either": [1, 2, 5], "serializ": [1, 5], "remoteobject": [1, 5], "await": [1, 2, 3, 4, 5, 6, 7], "valu": [1, 2, 3, 4, 5, 6], "idonli": [1, 5], "one": [1, 3, 4, 5, 6], "int": [1, 2, 3, 4, 5, 6], "depth": [1, 5, 6], "object": [1, 3, 4, 5], "complet": [1, 5, 7], "run": [1, 4, 5, 7], "exclus": [1, 5], "runtim": [1, 5, 6], "callfunctionon": [1, 5], "execute_script": [1, 3, 5, 6], "synchron": [1, 3, 5, 6], "descript": [1, 3, 5, 6], "execute_async_script": [1, 3, 5, 6], "asynchron": [1, 3, 5, 6], "i": [1, 2, 3, 4, 5, 6, 7], "recommend": [1, 3, 6], "except": [1, 3, 6], "correctli": [1, 3, 6], "eval_async": [1, 3, 5, 6], "resolv": [1, 3, 5, 6], "length": [1, 3, 5, 6], "1": [1, 3, 4, 5, 6], "re": [1, 3, 5, 6], "fetch": [1, 3, 4, 5, 6], "httpbin": [1, 3, 4, 5, 6], "mind": [1, 3, 4, 5, 6], "cor": [1, 3, 5, 6], "current_url": [1, 3, 5], "page_sourc": [1, 3, 5], "html": [1, 3, 5, 6, 7], "close": [1, 3, 5, 6], "focu": [1, 3, 6], "focus": [1, 6], "quit": [1, 3], "clean_dir": 1, "clean": [1, 2], "out": 1, "user": [1, 2, 4, 5, 6], "data": [1, 2, 3, 4, 5, 6], "dir": [1, 2], "current_target_info": 1, "current_window_handl": [1, 3], "deprec": [1, 2], "remov": [1, 2, 3, 5, 6], "current_window_id": 1, "windowid": 1, "belong": 1, "window_handl": [1, 3], "aren": [1, 7], "order": [1, 4], "posit": [1, 3], "do": [1, 6], "reli": 1, "index": 1, "iter": [1, 3, 4, 5], "them": 1, "new_window": [1, 3], "type_hint": [1, 3], "activ": [1, 3], "should": [1, 2, 3, 4, 5], "explicitli": 1, "new_target": 1, "set_window_st": 1, "state": 1, "normal": 1, "minim": [1, 3], "maxim": [1, 3, 5], "fullscreen": 1, "normalize_window": 1, "size": [1, 6], "maximize_window": [1, 3], "fullscreen_window": [1, 3], "enter": [1, 4], "minimize_window": [1, 3], "rtype": [1, 2], "throttl": 1, "print_pag": [1, 3, 5], "print": [1, 2, 4, 5, 7], "switch_to": [1, 3, 7], "switchto": [1, 3, 7], "back": [1, 3, 5], "goe": [1, 3, 5], "step": [1, 3, 5], "backward": [1, 3, 5], "histori": [1, 3, 5], "ha": [1, 2, 4, 5, 6], "forward": [1, 3, 5], "refresh": [1, 3, 5], "get_cooki": [1, 3, 5], "cooki": [1, 3, 5], "singl": [1, 2, 3, 4, 5], "found": [1, 3, 5], "delete_cooki": [1, 3, 5], "domain": [1, 3, 5], "delet": [1, 3, 5], "given": [1, 3, 5, 6], "delete_all_cooki": [1, 3, 5], "add_cooki": [1, 3, 5], "cookie_dict": [1, 3, 5], "add": [1, 2, 3, 5], "network": [1, 3, 4, 5], "cookieparam": [1, 5], "static": [1, 4], "sleep": [1, 7], "time_to_wait": 1, "sync": 1, "find_el": [1, 3, 5, 6, 7], "find": [1, 3, 5, 6], "element": [1, 3, 5, 6, 7], "locat": [1, 3, 5, 6, 7], "By": [1, 2, 3, 4, 5, 6, 7], "actual": [1, 3, 5, 6], "queri": [1, 3, 5, 6], "how": [1, 3, 5, 6], "long": [1, 3, 4, 5, 6], "exist": [1, 3, 5, 6, 7], "multipl": [1, 2, 3, 5, 6], "search_el": [1, 3, 5], "similarli": 1, "ctrl": 1, "f": 1, "devtool": [1, 2, 3, 5], "consol": 1, "plain": [1, 3, 5], "text": [1, 3, 4, 5, 6, 7], "get_screenshot_as_fil": [1, 3, 5], "filenam": [1, 3, 5, 6], "save": [1, 2, 3, 5, 6], "screenshot": [1, 3, 5, 6], "png": [1, 3, 5, 6], "imag": [1, 3, 5, 6], "wish": [1, 3, 5], "your": [1, 3, 4, 5], "end": [1, 3, 5], "save_screenshot": [1, 3, 5], "func": 1, "get_screenshot_as_png": [1, 3, 5], "binari": [1, 2, 3, 4, 5, 6], "byte": [1, 3, 4, 5, 6], "get_screenshot_as_base64": [1, 3, 5], "base64": [1, 3, 4, 5, 6], "encod": [1, 3, 4, 5], "embed": [1, 3, 5], "set_window_s": [1, 3], "width": [1, 3], "height": [1, 3], "unless": [1, 2], "windowhandl": [1, 3], "pixel": [1, 3], "get_window_s": [1, 3], "code": [1, 4], "block": [1, 4], "1280": 1, "720": 1, "set_window_posit": [1, 3], "x": [1, 3], "y": [1, 3], "coordin": [1, 3, 6], "get_window_posit": [1, 3], "0": [1, 3, 4, 6, 7], "get_window_rect": [1, 3], "windowst": 1, "set_window_rect": [1, 3], "get_network_condit": [1, 3, 5], "emul": [1, 3, 5], "latenc": [1, 3, 5], "4": [1, 3, 5], "download_throughput": [1, 3, 5], "upload_throughput": [1, 3, 5], "offlin": [1, 3, 5], "set_network_condit": [1, 3, 5], "connection_typ": [1, 3, 5], "addit": [1, 3, 5], "m": [1, 3, 5, 7], "throughput": [1, 3, 5], "500": [1, 3, 5], "1024": [1, 3, 5], "exampl": [1, 2, 3, 5, 7], "cellular2g": 1, "cellular3g": 1, "cellular4g": 1, "bluetooth": 1, "ethernet": 1, "wifi": [1, 3, 5], "wimax": 1, "both": [1, 3, 5], "upload": [1, 3, 5], "delete_network_condit": [1, 3, 5], "reset": [1, 3, 5], "set_permiss": [1, 3], "applic": [1, 3, 4], "permiss": [1, 3], "item": [1, 3], "prompt": 1, "appli": [1, 2], "clipboard": [1, 3], "read": [1, 3], "set_proxi": 1, "proxy_config": 1, "dynam": [1, 5], "mode": [1, 5], "fixed_serv": 1, "rule": 1, "proxyforhttp": 1, "scheme": [1, 4], "host": [1, 2, 5], "port": [1, 2], "proxyforftp": 1, "fallbackproxi": 1, "bypasslist": 1, "local": 1, "authentif": [1, 4], "set_auth": 1, "set_single_proxi": 1, "bypass_list": 1, "user1": [1, 2], "passwrd1": [1, 2], "5001": [1, 2], "when": [1, 2, 3, 4, 5], "been": [1, 2, 4], "driverless": [1, 2], "socks5": [1, 2], "authent": [1, 2, 4, 5], "1309413": [1, 2], "clear_proxi": 1, "clear": [1, 6], "usernam": [1, 4], "password": [1, 4, 5], "host_with_port": 1, "format": [1, 2, 6], "clear_auth": 1, "auth": [1, 4], "wait_for_cdp": [1, 3, 5, 7], "event": [1, 3, 4, 5, 6], "even": [], "windowopen": [], "add_cdp_listen": [1, 3, 5], "callback": [1, 3, 4, 5], "listen": [1, 3, 5, 6], "callabl": [4, 5], "accept": [2, 5, 6, 7], "remove_cdp_listen": [1, 3, 5], "execute_cdp_cmd": [1, 3, 5], "cmd": [1, 3, 5], "cmd_arg": [1, 3, 5], "10": [1, 3, 5, 7], "command": [1, 3, 5, 6], "result": 5, "follow": 5, "link": [4, 5], "chromedevtool": 5, "github": [5, 7], "io": 5, "A": [1, 3, 5, 6], "empti": 5, "getresponsebodi": 5, "requestid": [4, 5], "base64encod": 5, "bodi": [4, 5, 7], "respons": [4, 5], "get_sink": [1, 3, 5], "sink": [1, 3, 5], "avail": [1, 3, 5], "cast": [1, 3, 5], "get_issue_messag": [1, 3, 5], "error": [1, 3, 4, 5], "messag": [1, 3, 5], "issu": [1, 3, 5], "session": [1, 2, 3, 5], "set_sink_to_us": [1, 3, 5], "sink_nam": [1, 3, 5], "its": [1, 3, 5], "receiv": [1, 3, 5], "start_desktop_mirror": [1, 3, 5], "desktop": [1, 3, 5], "mirror": [1, 3, 5], "start_tab_mirror": [1, 3, 5], "stop_cast": [1, 3, 5], "stop": [1, 3, 5], "loop": [1, 3, 5, 6], "alert": [1, 3, 7], "get_alert": 1, "5": [1, 2, 3, 5, 6, 7], "default_cont": [1, 3], "frame_refer": 1, "integ": 1, "repres": [1, 5, 6], "frame_nam": [1, 3], "union": [1, 2, 4, 6], "what": 1, "kind": 1, "parent_fram": [1, 3], "parent": [1, 6], "If": [1, 3, 6], "top": [1, 3], "level": [1, 3, 6], "brows": [1, 3], "remain": 1, "unchang": 1, "window_id": 1, "window_nam": 1, "main": [1, 3, 4, 7], "chrome": [2, 3, 4, 5, 6, 7], "add_argu": 2, "launch": 2, "pref": 2, "prefer": 2, "update_pref": 2, "updat": 2, "dot": 2, "user_data_dir": 2, "temporarili": 2, "temp": 2, "_dir": 2, "o": 2, "getcwd": 2, "isdir": 2, "mkdir": 2, "webdriv": [2, 3, 4, 7], "plugin": 2, "always_open_pdf_extern": 2, "download_data": 2, "absolut": [2, 6], "headless": 2, "startup_url": 2, "binary_loc": 2, "env": 2, "subprocess": 2, "popen": 2, "environ": 2, "add_extens": 2, "zip": 2, "crx": 2, "etc": 2, "extract": 2, "debugger_address": 2, "address": 2, "remot": 2, "instanc": [2, 5, 6], "make": [2, 3, 4, 5], "well": [2, 3], "single_proxi": 2, "auto_clean_dir": 2, "automat": [2, 3, 4, 6], "enable_mobil": 2, "android_packag": 2, "android": 2, "android_act": 2, "device_seri": 2, "enabl": [2, 6], "mobil": 2, "packag": 2, "Not": [1, 2, 3, 6], "implement": [2, 6], "accept_insecure_cert": 2, "insecur": 2, "certif": 2, "notimpl": [2, 6], "ignore_local_proxy_environment_vari": 2, "call": [2, 6], "http_proxi": 2, "https_proxi": 2, "being": [2, 6], "pick": 2, "up": 2, "add_experimental_opt": 2, "experiment": 2, "method": [2, 3, 4, 5, 6, 7], "is_incognito": 3, "1048576": [3, 5], "drive": [3, 5], "chromedriv": [3, 5], "start_sess": 3, "desir": 3, "capabl": 3, "behavior": [3, 4], "click": [3, 6, 7], "const": 3, "accord": 3, "obj_id": [3, 6], "docs_sourc": 3, "start_monoton": 3, "everi": [3, 6], "associ": 3, "hint": 3, "select": [3, 6], "invok": 3, "manag": 3, "full": [3, 5], "screen": 3, "oper": [3, 6], "take": [3, 4, 5, 6], "best": [3, 5], "effort": [3, 5], "provid": [3, 4, 5], "contain": 3, "active_el": 3, "dictionari": [3, 5, 6], "correspond": [3, 4, 5], "visibl": [3, 5, 6], "my_cooki": 3, "scope": [3, 5], "requir": [3, 4], "kei": [3, 4, 5], "secur": 3, "httponli": 3, "expiri": 3, "samesit": 3, "foo": [3, 5], "bar": 3, "strict": [3, 4], "search": [3, 5], "ioerror": [3, 5], "els": [3, 4, 5, 6], "resizeto": 3, "800": 3, "600": 3, "moveto": 3, "w3c": 3, "compat": 3, "100": 3, "200": [3, 4], "network_condit": [3, 5], "condit": [3, 5], "note": [3, 4, 5], "network_interceptor": 4, "requestdoneexcept": 4, "authalreadyhandledexcept": 4, "requestpattern": 4, "enumer": 4, "is_remot": 5, "start_socket": 5, "get_histori": 5, "getnavigationhistori": 5, "navigationhistori": 5, "set_sourc": [5, 6], "15": [5, 7], "outerhtml": [5, 6], "dom": [5, 6], "try": 5, "might": [4, 5, 6], "fail": [4, 5, 6], "reload": 5, "target_info": 5, "target_gett": 5, "info": 5, "itself": 5, "attach": [5, 6], "client": 5, "opener_id": 5, "open": 5, "can_access_open": 5, "opener_frame_id": 5, "browser_context_id": 5, "browsercontextid": 5, "subtyp": 5, "detail": [5, 6], "mai": 5, "portal": 5, "prerend": 5, "frame_id": [4, 6], "isolated_exec_id": 6, "node_id": 6, "backend_node_id": 6, "is_ifram": 6, "interest": 6, "interact": 6, "perform": 6, "through": 6, "interfac": 6, "fresh": 6, "check": 6, "still": [4, 6], "valid": 6, "essenti": 6, "determin": 6, "staleelementreferenceexcept": 6, "thrown": 6, "futur": 6, "remoteobjectid": 6, "document_url": 6, "idx": 6, "get_listen": 6, "domdebugg": 6, "geteventlisten": 6, "get_properti": 6, "elem": [6, 7], "tagnam": 6, "textcont": 6, "entri": 6, "highlight": 6, "disabl": 6, "fade": 6, "is_click": 6, "listener_depth": 6, "button": 6, "input": 6, "textarea": 6, "video": 6, "map": 6, "wise": 6, "mousedown": 6, "mouseup": 6, "visible_timeout": 6, "spread_a": 6, "spread_b": 6, "bias_a": 6, "bias_b": 6, "border": 6, "05": 6, "scroll_to": 6, "move_to": [6, 7], "ensure_click": 6, "abl": 6, "comput": 6, "box": 6, "model": 6, "spread": 6, "over": [1, 3, 4, 5, 6], "b": 6, "bia": 6, "minimum": 6, "toward": 6, "edg": 6, "rel": 6, "random": 6, "point": 6, "outsid": 6, "scroll": 6, "move": 6, "mous": 6, "clickabl": 6, "reliabl": 6, "equival": 6, "6": 6, "std": 6, "99": 6, "7": 6, "write": 6, "insert": 6, "send": 6, "set_fil": 6, "send_kei": 6, "mid_loc": 6, "probabl": 6, "middl": 6, "submit": 6, "form": 6, "detect": [4, 6], "It": 6, "possibl": [4, 6], "get_dom_attribut": 6, "attribut": 6, "unlik": 6, "get_attribut": 6, "declar": 6, "markup": 6, "retriev": 6, "text_length": 6, "target_el": 6, "is_select": 6, "checkbox": 6, "radio": 6, "is_en": 6, "shadow_root": 6, "shadowroot": 6, "doe": [4, 6], "shadow": 6, "is_displai": 6, "location_once_scrolled_into_view": 6, "rect": 6, "properli": [6, 7], "value_of_css_properti": 6, "property_nam": 6, "render": 6, "canva": 6, "box_model": 6, "boxmodel": 6, "aria_rol": 6, "aria": 6, "role": 6, "accessible_nam": 6, "screenshot_as_base64": 6, "screenshot_as_png": 6, "chromeopt": 7, "request": [1, 3, 5, 7], "intercept": 7, "networkinterceptor": 4, "on_request": 4, "on_respons": 4, "on_auth": 4, "pattern": 4, "intercept_auth": 4, "bypass_service_work": 4, "__init__": 4, "interceptedrequest": 4, "onrequest": 4, "onrespons": 4, "onauth": 4, "bypass": 4, "servic": 4, "worker": 4, "patter": 4, "__aiter__": 4, "asynciter": [1, 3, 4, 5], "interceptedauth": 4, "ware": [], "virtual": 4, "zero": 4, "asyncio": [4, 7], "ensure_futur": 4, "where": 4, "warn": [], "import": [4, 7], "requeststag": 4, "def": [4, 7], "post": [4, 5], "continue_request": 4, "basic": 4, "intercept_respons": 4, "anyrequest": 4, "interceptor": 4, "isinst": 4, "would": 4, "continue_auth": 4, "elif": 4, "stage": 4, "break": 4, "python": 7, "requestpaus": 4, "authrequir": 4, "get_cdp_event_it": [1, 3, 5], "socket": 5, "singlecdpsocket": 5, "framenavig": 5, "now": 5, "chromeinst": 5, "whole": 5, "usual": 5, "unfortun": 5, "downloads_dir_for_context": 5, "is_redirect": 4, "redirect": [4, 5], "response_status_cod": 4, "301": 4, "302": 4, "303": 4, "307": 4, "308": 4, "resource_typ": 4, "resourcetyp": 4, "network_id": 4, "requestwillbes": 4, "had": 4, "fire": 4, "response_error_reason": 4, "reason": 4, "response_head": 4, "header": [4, 5], "statu": [4, 5], "response_status_text": 4, "redirected_id": 4, "bypass_brows": 4, "allow_redirect": 4, "proxy_auth": 4, "extern": 4, "tl": 4, "fingerprint": 4, "accordingli": 4, "technic": 4, "post_data": 4, "continu": 4, "fulfil": 4, "arrai": 4, "observ": 4, "overrid": 4, "continue_respons": 4, "binary_response_head": 4, "response_cod": 4, "response_phras": 4, "separ": 4, "seri": 4, "pair": 4, "treat": 4, "phrase": 4, "ok": [4, 5], "resum": 4, "rais": 4, "cancel": 4, "done": 4, "fail_request": 4, "error_reason": 4, "abort": 4, "timedout": 4, "accessdeni": 4, "connectionclos": 4, "connectionreset": 4, "connectionrefus": 4, "connectionabort": 4, "connectionfail": 4, "namenotresolv": 4, "internetdisconnect": 4, "addressunreach": 4, "blockedbycli": 4, "blockedbyrespons": 4, "auth_challeng": 4, "authchalleng": 4, "cancelauth": 4, "providecredenti": 4, "challeng": 4, "digest": 4, "realm": 4, "pleas": 4, "fragment": 4, "url_frag": 4, "hash": 4, "present": 4, "has_post_data": 4, "postdata": 4, "omit": 4, "flag": 4, "too": 4, "post_data_entri": 4, "postdataentri": 4, "mixed_content_typ": 4, "blockabl": 4, "mix": 4, "content": 4, "initial_prior": 4, "verylow": 4, "low": 4, "medium": 4, "high": [4, 5], "veryhigh": 4, "prioriti": [4, 5], "sent": 4, "referrer_polici": [4, 5], "unsaf": 4, "downgrad": 4, "polici": 4, "defin": 4, "tr": 4, "is_link_preload": 4, "via": 4, "preload": 4, "trust_token_param": 4, "trusttokenparam": 4, "is_same_sit": 4, "consid": 4, "urlpattern": 4, "anyrespons": 4, "url_pattern": 4, "request_stag": 4, "nowsecur": 7, "nl": 7, "relax": 7, "domcontenteventfir": 7, "div": 7, "p": 7, "kaliiiiiiiiii": 7, "selenium": 7, "pip": 7, "upgrad": 7, "kwarg": [1, 3], "j": [1, 3, 5], "xhr": [1, 3, 5], "xmlhttprequest": [1, 3, 5], "credenti": 5, "cach": 5, "integr": 5, "keepal": 5, "20": 5, "mozilla": 5, "en": 5, "u": 5, "api": 5, "with_credenti": 5, "mime_typ": 5, "xml": [], "extra_head": 5, "smth": 5, "status_cod": 5, "status_text": 5, "put": 5, "includ": 5, "pars": 5, "extra": 5, "responsetext": 5, "responsetyp": 5, "responseurl": 5, "responsexml": 5, "statustext": 5, "responsehead": 5}, "objects": {"selenium_driverless.scripts.network_interceptor": [[4, 0, 1, "", "AuthAlreadyHandledException"], [4, 0, 1, "", "AuthChallenge"], [4, 0, 1, "", "InterceptedAuth"], [4, 0, 1, "", "InterceptedRequest"], [4, 0, 1, "", "NetworkInterceptor"], [4, 0, 1, "", "Request"], [4, 0, 1, "", "RequestDoneException"], [4, 0, 1, "", "RequestPattern"], [4, 0, 1, "", "RequestStages"]], "selenium_driverless.scripts.network_interceptor.AuthAlreadyHandledException": [[4, 1, 1, "", "request"]], "selenium_driverless.scripts.network_interceptor.AuthChallenge": [[4, 1, 1, "", "origin"], [4, 1, 1, "", "params"], [4, 1, 1, "", "realm"], [4, 1, 1, "", "scheme"], [4, 1, 1, "", "source"], [4, 1, 1, "", "target"]], "selenium_driverless.scripts.network_interceptor.InterceptedAuth": [[4, 1, 1, "", "auth_challenge"], [4, 2, 1, "", "cancel"], [4, 2, 1, "", "continue_auth"], [4, 1, 1, "", "frame_id"], [4, 1, 1, "", "id"], [4, 1, 1, "", "params"], [4, 1, 1, "", "request"], [4, 1, 1, "", "resource_type"], [4, 2, 1, "", "resume"], [4, 1, 1, "", "target"]], "selenium_driverless.scripts.network_interceptor.InterceptedRequest": [[4, 1, 1, "", "body"], [4, 2, 1, "", "bypass_browser"], [4, 2, 1, "", "continue_request"], [4, 2, 1, "", "continue_response"], [4, 2, 1, "", "fail_request"], [4, 1, 1, "", "frame_id"], [4, 2, 1, "", "fulfill"], [4, 1, 1, "", "id"], [4, 1, 1, "", "is_redirect"], [4, 1, 1, "", "network_id"], [4, 1, 1, "", "params"], [4, 1, 1, "", "redirected_id"], [4, 1, 1, "", "request"], [4, 1, 1, "", "resource_type"], [4, 1, 1, "", "response_error_reason"], [4, 1, 1, "", "response_headers"], [4, 1, 1, "", "response_status_code"], [4, 1, 1, "", "response_status_text"], [4, 2, 1, "", "resume"], [4, 1, 1, "", "stage"], [4, 1, 1, "", "target"]], "selenium_driverless.scripts.network_interceptor.NetworkInterceptor": [[4, 2, 1, "", "__aiter__"], [4, 2, 1, "", "__init__"], [4, 1, 1, "", "patterns"], [4, 1, 1, "", "target"]], "selenium_driverless.scripts.network_interceptor.Request": [[4, 1, 1, "", "has_post_data"], [4, 1, 1, "", "headers"], [4, 1, 1, "", "initial_priority"], [4, 1, 1, "", "is_link_preload"], [4, 1, 1, "", "is_same_site"], [4, 1, 1, "", "method"], [4, 1, 1, "", "mixed_content_type"], [4, 1, 1, "", "params"], [4, 1, 1, "", "post_data"], [4, 1, 1, "", "post_data_entries"], [4, 1, 1, "", "referrer_policy"], [4, 1, 1, "", "target"], [4, 1, 1, "", "trust_token_params"], [4, 1, 1, "", "url"], [4, 1, 1, "", "url_fragment"]], "selenium_driverless.scripts.network_interceptor.RequestDoneException": [[4, 1, 1, "", "request"]], "selenium_driverless.scripts.network_interceptor.RequestPattern": [[4, 3, 1, "", "AnyRequest"], [4, 3, 1, "", "AnyResponse"], [4, 2, 1, "", "new"]], "selenium_driverless.scripts.network_interceptor.RequestStages": [[4, 3, 1, "", "Request"], [4, 3, 1, "", "Response"]], "selenium_driverless.scripts.switch_to": [[1, 0, 1, "", "SwitchTo"]], "selenium_driverless.scripts.switch_to.SwitchTo": [[1, 1, 1, "", "alert"], [1, 2, 1, "", "default_content"], [1, 2, 1, "", "frame"], [1, 2, 1, "", "get_alert"], [1, 2, 1, "", "new_window"], [1, 2, 1, "", "parent_frame"], [1, 2, 1, "", "target"], [1, 2, 1, "", "window"]], "selenium_driverless.types.base_target": [[5, 0, 1, "", "BaseTarget"]], "selenium_driverless.types.base_target.BaseTarget": [[5, 2, 1, "", "add_cdp_listener"], [5, 2, 1, "", "downloads_dir_for_context"], [5, 2, 1, "", "execute_cdp_cmd"], [5, 2, 1, "", "get_cdp_event_iter"], [5, 2, 1, "", "remove_cdp_listener"], [5, 1, 1, "", "socket"], [5, 2, 1, "", "wait_for_cdp"]], "selenium_driverless.types.by": [[0, 0, 1, "", "By"]], "selenium_driverless.types.by.By": [[0, 3, 1, "", "CLASS_NAME"], [0, 3, 1, "", "CSS"], [0, 3, 1, "", "CSS_SELECTOR"], [0, 3, 1, "", "ID"], [0, 3, 1, "", "NAME"], [0, 3, 1, "", "TAG_NAME"], [0, 3, 1, "", "XPATH"]], "selenium_driverless.types.context": [[3, 0, 1, "", "Context"]], "selenium_driverless.types.context.Context": [[3, 2, 1, "", "add_cdp_listener"], [3, 2, 1, "", "add_cookie"], [3, 2, 1, "", "back"], [3, 2, 1, "", "close"], [3, 1, 1, "", "current_url"], [3, 1, 1, "", "current_window_handle"], [3, 2, 1, "", "delete_all_cookies"], [3, 2, 1, "", "delete_cookie"], [3, 2, 1, "", "delete_network_conditions"], [3, 1, 1, "", "downloads_dir"], [3, 2, 1, "", "eval_async"], [3, 2, 1, "", "execute_async_script"], [3, 2, 1, "", "execute_cdp_cmd"], [3, 2, 1, "", "execute_raw_script"], [3, 2, 1, "", "execute_script"], [3, 2, 1, "", "fetch"], [3, 2, 1, "", "find_element"], [3, 2, 1, "", "find_elements"], [3, 2, 1, "", "forward"], [3, 2, 1, "", "fullscreen_window"], [3, 2, 1, "", "get"], [3, 2, 1, "", "get_cdp_event_iter"], [3, 2, 1, "", "get_cookie"], [3, 2, 1, "", "get_cookies"], [3, 2, 1, "", "get_issue_message"], [3, 2, 1, "", "get_network_conditions"], [3, 2, 1, "", "get_screenshot_as_base64"], [3, 2, 1, "", "get_screenshot_as_file"], [3, 2, 1, "", "get_screenshot_as_png"], [3, 2, 1, "", "get_sinks"], [3, 2, 1, "", "get_window_position"], [3, 2, 1, "", "get_window_rect"], [3, 2, 1, "", "get_window_size"], [3, 2, 1, "", "maximize_window"], [3, 2, 1, "", "minimize_window"], [3, 2, 1, "", "new_window"], [3, 1, 1, "", "page_source"], [3, 2, 1, "", "print_page"], [3, 2, 1, "", "quit"], [3, 2, 1, "", "refresh"], [3, 2, 1, "", "remove_cdp_listener"], [3, 2, 1, "", "save_screenshot"], [3, 2, 1, "", "search_elements"], [3, 2, 1, "", "set_download_behaviour"], [3, 2, 1, "", "set_network_conditions"], [3, 2, 1, "", "set_permissions"], [3, 2, 1, "", "set_sink_to_use"], [3, 2, 1, "", "set_window_position"], [3, 2, 1, "", "set_window_rect"], [3, 2, 1, "", "set_window_size"], [3, 2, 1, "", "start_desktop_mirroring"], [3, 2, 1, "", "start_session"], [3, 2, 1, "", "start_tab_mirroring"], [3, 2, 1, "", "stop_casting"], [3, 1, 1, "", "switch_to"], [3, 1, 1, "", "title"], [3, 2, 1, "", "wait_download"], [3, 2, 1, "", "wait_for_cdp"], [3, 1, 1, "", "window_handles"], [3, 2, 1, "", "xhr"]], "selenium_driverless.types.options": [[2, 0, 1, "", "Options"]], "selenium_driverless.types.options.Options": [[2, 1, 1, "", "accept_insecure_certs"], [2, 2, 1, "", "add_argument"], [2, 2, 1, "", "add_arguments"], [2, 2, 1, "", "add_experimental_option"], [2, 2, 1, "", "add_extension"], [2, 1, 1, "", "arguments"], [2, 1, 1, "", "auto_clean_dirs"], [2, 1, 1, "", "binary_location"], [2, 1, 1, "", "debugger_address"], [2, 1, 1, "", "downloads_dir"], [2, 2, 1, "", "enable_mobile"], [2, 1, 1, "", "env"], [2, 1, 1, "", "headless"], [2, 2, 1, "", "ignore_local_proxy_environment_variables"], [2, 1, 1, "", "prefs"], [2, 1, 1, "", "single_proxy"], [2, 1, 1, "", "startup_url"], [2, 2, 1, "", "update_pref"], [2, 1, 1, "", "user_data_dir"]], "selenium_driverless.types.target": [[5, 0, 1, "", "Target"], [5, 0, 1, "", "TargetInfo"]], "selenium_driverless.types.target.Target": [[5, 2, 1, "", "add_cdp_listener"], [5, 2, 1, "", "add_cookie"], [5, 2, 1, "", "back"], [5, 2, 1, "", "close"], [5, 1, 1, "", "current_url"], [5, 2, 1, "", "delete_all_cookies"], [5, 2, 1, "", "delete_cookie"], [5, 2, 1, "", "delete_network_conditions"], [5, 2, 1, "", "eval_async"], [5, 2, 1, "", "execute_async_script"], [5, 2, 1, "", "execute_cdp_cmd"], [5, 2, 1, "", "execute_raw_script"], [5, 2, 1, "", "execute_script"], [5, 2, 1, "", "fetch"], [5, 2, 1, "", "find_element"], [5, 2, 1, "", "find_elements"], [5, 2, 1, "", "forward"], [5, 2, 1, "", "get"], [5, 2, 1, "", "get_cdp_event_iter"], [5, 2, 1, "", "get_cookie"], [5, 2, 1, "", "get_cookies"], [5, 2, 1, "", "get_history"], [5, 2, 1, "", "get_issue_message"], [5, 2, 1, "", "get_network_conditions"], [5, 2, 1, "", "get_screenshot_as_base64"], [5, 2, 1, "", "get_screenshot_as_file"], [5, 2, 1, "", "get_screenshot_as_png"], [5, 2, 1, "", "get_sinks"], [5, 1, 1, "", "page_source"], [5, 2, 1, "", "print_page"], [5, 2, 1, "", "refresh"], [5, 2, 1, "", "remove_cdp_listener"], [5, 2, 1, "", "save_screenshot"], [5, 2, 1, "", "search_elements"], [5, 2, 1, "", "set_network_conditions"], [5, 2, 1, "", "set_sink_to_use"], [5, 2, 1, "", "set_source"], [5, 1, 1, "", "socket"], [5, 2, 1, "", "start_desktop_mirroring"], [5, 2, 1, "", "start_tab_mirroring"], [5, 2, 1, "", "stop_casting"], [5, 1, 1, "", "title"], [5, 2, 1, "", "wait_download"], [5, 2, 1, "", "wait_for_cdp"], [5, 2, 1, "", "xhr"]], "selenium_driverless.types.target.TargetInfo": [[5, 1, 1, "", "Target"], [5, 1, 1, "", "attached"], [5, 1, 1, "", "browser_context_id"], [5, 1, 1, "", "can_access_opener"], [5, 1, 1, "", "id"], [5, 1, 1, "", "opener_frame_id"], [5, 1, 1, "", "opener_id"], [5, 1, 1, "", "subtype"]], "selenium_driverless.types.webelement": [[6, 0, 1, "", "WebElement"]], "selenium_driverless.types.webelement.WebElement": [[6, 1, 1, "", "accessible_name"], [6, 1, 1, "", "aria_role"], [6, 1, 1, "", "box_model"], [6, 2, 1, "", "clear"], [6, 2, 1, "", "click"], [6, 1, 1, "", "content_document"], [6, 1, 1, "", "document_url"], [6, 2, 1, "", "eval_async"], [6, 2, 1, "", "execute_async_script"], [6, 2, 1, "", "execute_script"], [6, 2, 1, "", "find_element"], [6, 2, 1, "", "find_elements"], [6, 2, 1, "", "focus"], [6, 2, 1, "", "get_attribute"], [6, 2, 1, "", "get_dom_attribute"], [6, 2, 1, "", "get_listeners"], [6, 2, 1, "", "get_property"], [6, 2, 1, "", "highlight"], [6, 2, 1, "", "is_clickable"], [6, 2, 1, "", "is_displayed"], [6, 2, 1, "", "is_enabled"], [6, 2, 1, "", "is_selected"], [6, 1, 1, "", "location"], [6, 1, 1, "", "location_once_scrolled_into_view"], [6, 2, 1, "", "mid_location"], [6, 1, 1, "", "obj_id"], [6, 1, 1, "", "parent"], [6, 1, 1, "", "rect"], [6, 2, 1, "", "remove"], [6, 2, 1, "", "screenshot"], [6, 1, 1, "", "screenshot_as_base64"], [6, 1, 1, "", "screenshot_as_png"], [6, 2, 1, "", "scroll_to"], [6, 2, 1, "", "send_keys"], [6, 2, 1, "", "set_file"], [6, 2, 1, "", "set_files"], [6, 2, 1, "", "set_source"], [6, 1, 1, "", "shadow_root"], [6, 1, 1, "", "size"], [6, 1, 1, "", "source"], [6, 2, 1, "", "submit"], [6, 1, 1, "", "tag_name"], [6, 1, 1, "", "text"], [6, 1, 1, "", "value"], [6, 2, 1, "", "value_of_css_property"], [6, 2, 1, "", "write"]], "selenium_driverless.webdriver": [[1, 0, 1, "", "Chrome"]], "selenium_driverless.webdriver.Chrome": [[1, 2, 1, "", "add_cdp_listener"], [1, 2, 1, "", "add_cookie"], [1, 2, 1, "", "back"], [1, 1, 1, "", "base_context"], [1, 1, 1, "", "base_target"], [1, 2, 1, "", "clear_auth"], [1, 2, 1, "", "clear_proxy"], [1, 2, 1, "", "close"], [1, 1, 1, "", "contexts"], [1, 1, 1, "", "current_context"], [1, 1, 1, "", "current_pointer"], [1, 1, 1, "", "current_target"], [1, 1, 1, "", "current_target_info"], [1, 1, 1, "", "current_url"], [1, 1, 1, "", "current_window_handle"], [1, 1, 1, "", "current_window_id"], [1, 2, 1, "", "delete_all_cookies"], [1, 2, 1, "", "delete_cookie"], [1, 2, 1, "", "delete_network_conditions"], [1, 1, 1, "", "downloads_dir"], [1, 2, 1, "", "ensure_extensions_incognito_allowed"], [1, 2, 1, "", "eval_async"], [1, 2, 1, "", "execute_async_script"], [1, 2, 1, "", "execute_cdp_cmd"], [1, 2, 1, "", "execute_raw_script"], [1, 2, 1, "", "execute_script"], [1, 2, 1, "", "fetch"], [1, 2, 1, "", "find_element"], [1, 2, 1, "", "find_elements"], [1, 2, 1, "", "focus"], [1, 2, 1, "", "forward"], [1, 1, 1, "", "frame_tree"], [1, 2, 1, "", "fullscreen_window"], [1, 2, 1, "", "get"], [1, 2, 1, "", "get_cdp_event_iter"], [1, 2, 1, "", "get_cookie"], [1, 2, 1, "", "get_cookies"], [1, 2, 1, "", "get_issue_message"], [1, 2, 1, "", "get_network_conditions"], [1, 2, 1, "", "get_screenshot_as_base64"], [1, 2, 1, "", "get_screenshot_as_file"], [1, 2, 1, "", "get_screenshot_as_png"], [1, 2, 1, "", "get_sinks"], [1, 2, 1, "", "get_target"], [1, 2, 1, "", "get_target_for_iframe"], [1, 2, 1, "", "get_targets"], [1, 2, 1, "", "get_targets_for_iframes"], [1, 2, 1, "", "get_window_position"], [1, 2, 1, "", "get_window_rect"], [1, 2, 1, "", "get_window_size"], [1, 2, 1, "", "maximize_window"], [1, 2, 1, "", "minimize_window"], [1, 1, 1, "", "mv3_extension"], [1, 2, 1, "", "new_context"], [1, 2, 1, "", "new_window"], [1, 2, 1, "", "normalize_window"], [1, 1, 1, "", "page_source"], [1, 2, 1, "", "print_page"], [1, 2, 1, "", "quit"], [1, 2, 1, "", "refresh"], [1, 2, 1, "", "remove_cdp_listener"], [1, 2, 1, "", "save_screenshot"], [1, 2, 1, "", "search_elements"], [1, 2, 1, "", "set_auth"], [1, 2, 1, "", "set_download_behaviour"], [1, 2, 1, "", "set_network_conditions"], [1, 2, 1, "", "set_permissions"], [1, 2, 1, "", "set_proxy"], [1, 2, 1, "", "set_single_proxy"], [1, 2, 1, "", "set_sink_to_use"], [1, 2, 1, "", "set_window_position"], [1, 2, 1, "", "set_window_rect"], [1, 2, 1, "", "set_window_size"], [1, 2, 1, "", "set_window_state"], [1, 2, 1, "", "sleep"], [1, 2, 1, "", "start_desktop_mirroring"], [1, 2, 1, "", "start_tab_mirroring"], [1, 2, 1, "", "stop_casting"], [1, 1, 1, "", "switch_to"], [1, 1, 1, "", "targets"], [1, 1, 1, "", "title"], [1, 2, 1, "", "wait_download"], [1, 2, 1, "", "wait_for_cdp"], [1, 1, 1, "", "window_handles"], [1, 2, 1, "", "xhr"]]}, "objtypes": {"0": "py:class", "1": "py:property", "2": "py:method", "3": "py:attribute"}, "objnames": {"0": ["py", "class", "Python class"], "1": ["py", "property", "Python property"], "2": ["py", "method", "Python method"], "3": ["py", "attribute", "Python attribute"]}, "titleterms": {"By": 0, "element": 0, "locat": 0, "webdriv": 1, "chrome": 1, "chromeopt": 2, "context": 3, "request": 4, "intercept": 4, "target": 5, "webel": 6, "document": 7, "selenium": [], "driverless": 7, "indic": [], "tabl": [], "api": [4, 7], "exampl": 4, "script": 4, "usag": 7, "instal": 7, "sourc": 7}, "envversion": {"sphinx.domains.c": 3, "sphinx.domains.changeset": 1, "sphinx.domains.citation": 1, "sphinx.domains.cpp": 9, "sphinx.domains.index": 1, "sphinx.domains.javascript": 3, "sphinx.domains.math": 2, "sphinx.domains.python": 4, "sphinx.domains.rst": 2, "sphinx.domains.std": 2, "sphinx.ext.viewcode": 1, "sphinx": 60}, "alltitles": {"By Element Locator": [[0, "by-element-locator"]], "ChromeOptions": [[2, "chromeoptions"]], "WebElement": [[6, "webelement"]], "Documentation of Driverless": [[7, "documentation-of-driverless"]], "Installation": [[7, "installation"]], "Usage": [[7, "usage"]], "API": [[7, "api"], [4, "api"]], "Source": [[7, "source"]], "webdriver.Chrome": [[1, "webdriver-chrome"]], "Context": [[3, "context"]], "Request-Interception": [[4, "request-interception"]], "Example Script": [[4, "example-script"]], "Target": [[5, "target"]]}, "indexentries": {"basetarget (class in selenium_driverless.types.base_target)": [[5, "selenium_driverless.types.base_target.BaseTarget"]], "target (class in selenium_driverless.types.target)": [[5, "selenium_driverless.types.target.Target"]], "target (selenium_driverless.types.target.targetinfo property)": [[5, "selenium_driverless.types.target.TargetInfo.Target"]], "targetinfo (class in selenium_driverless.types.target)": [[5, "selenium_driverless.types.target.TargetInfo"]], "add_cdp_listener() (selenium_driverless.types.base_target.basetarget method)": [[5, "selenium_driverless.types.base_target.BaseTarget.add_cdp_listener"]], "add_cdp_listener() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.add_cdp_listener"]], "add_cookie() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.add_cookie"]], "attached (selenium_driverless.types.target.targetinfo property)": [[5, "selenium_driverless.types.target.TargetInfo.attached"]], "back() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.back"]], "browser_context_id (selenium_driverless.types.target.targetinfo property)": [[5, "selenium_driverless.types.target.TargetInfo.browser_context_id"]], "can_access_opener (selenium_driverless.types.target.targetinfo property)": [[5, "selenium_driverless.types.target.TargetInfo.can_access_opener"]], "close() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.close"]], "current_url (selenium_driverless.types.target.target property)": [[5, "selenium_driverless.types.target.Target.current_url"]], "delete_all_cookies() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.delete_all_cookies"]], "delete_cookie() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.delete_cookie"]], "delete_network_conditions() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.delete_network_conditions"]], "downloads_dir_for_context() (selenium_driverless.types.base_target.basetarget method)": [[5, "selenium_driverless.types.base_target.BaseTarget.downloads_dir_for_context"]], "eval_async() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.eval_async"]], "execute_async_script() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.execute_async_script"]], "execute_cdp_cmd() (selenium_driverless.types.base_target.basetarget method)": [[5, "selenium_driverless.types.base_target.BaseTarget.execute_cdp_cmd"]], "execute_cdp_cmd() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.execute_cdp_cmd"]], "execute_raw_script() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.execute_raw_script"]], "execute_script() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.execute_script"]], "fetch() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.fetch"]], "find_element() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.find_element"]], "find_elements() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.find_elements"]], "forward() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.forward"]], "get() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.get"]], "get_cdp_event_iter() (selenium_driverless.types.base_target.basetarget method)": [[5, "selenium_driverless.types.base_target.BaseTarget.get_cdp_event_iter"]], "get_cdp_event_iter() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.get_cdp_event_iter"]], "get_cookie() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.get_cookie"]], "get_cookies() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.get_cookies"]], "get_history() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.get_history"]], "get_issue_message() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.get_issue_message"]], "get_network_conditions() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.get_network_conditions"]], "get_screenshot_as_base64() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.get_screenshot_as_base64"]], "get_screenshot_as_file() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.get_screenshot_as_file"]], "get_screenshot_as_png() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.get_screenshot_as_png"]], "get_sinks() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.get_sinks"]], "id (selenium_driverless.types.target.targetinfo property)": [[5, "selenium_driverless.types.target.TargetInfo.id"]], "opener_frame_id (selenium_driverless.types.target.targetinfo property)": [[5, "selenium_driverless.types.target.TargetInfo.opener_frame_id"]], "opener_id (selenium_driverless.types.target.targetinfo property)": [[5, "selenium_driverless.types.target.TargetInfo.opener_id"]], "page_source (selenium_driverless.types.target.target property)": [[5, "selenium_driverless.types.target.Target.page_source"]], "print_page() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.print_page"]], "refresh() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.refresh"]], "remove_cdp_listener() (selenium_driverless.types.base_target.basetarget method)": [[5, "selenium_driverless.types.base_target.BaseTarget.remove_cdp_listener"]], "remove_cdp_listener() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.remove_cdp_listener"]], "save_screenshot() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.save_screenshot"]], "search_elements() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.search_elements"]], "set_network_conditions() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.set_network_conditions"]], "set_sink_to_use() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.set_sink_to_use"]], "set_source() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.set_source"]], "socket (selenium_driverless.types.base_target.basetarget property)": [[5, "selenium_driverless.types.base_target.BaseTarget.socket"]], "socket (selenium_driverless.types.target.target property)": [[5, "selenium_driverless.types.target.Target.socket"]], "start_desktop_mirroring() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.start_desktop_mirroring"]], "start_tab_mirroring() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.start_tab_mirroring"]], "stop_casting() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.stop_casting"]], "subtype (selenium_driverless.types.target.targetinfo property)": [[5, "selenium_driverless.types.target.TargetInfo.subtype"]], "title (selenium_driverless.types.target.target property)": [[5, "selenium_driverless.types.target.Target.title"]], "wait_download() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.wait_download"]], "wait_for_cdp() (selenium_driverless.types.base_target.basetarget method)": [[5, "selenium_driverless.types.base_target.BaseTarget.wait_for_cdp"]], "wait_for_cdp() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.wait_for_cdp"]], "xhr() (selenium_driverless.types.target.target method)": [[5, "selenium_driverless.types.target.Target.xhr"]]}})
